refactor(design): D38 split generator into Spec/UI/Iterate stages
Extract gen_spec, gen_ui, and gen_iterate from generator.py; keep
generator as orchestration facade with re-exports. Retarget design
test monkeypatches to defining modules.
