+++ memory/PROGRESS.md
@@ -7,9 +7,15 @@
 
 - **阶段：** **v1.3.0** 已发；主线 D8–D13 ✅；扩展/MCP **D14–D52 Agent 代 PM ✅**；Desktop E2E ✅；Design D36 仍 ❌
 - **Git / PM 索引：** [`runs/verification/pm-acceptance/AGENT-PM-2026-07-25.md`](../runs/verification/pm-acceptance/AGENT-PM-2026-07-25.md)
-- **下次优先：** D38 拆 `design/generator.py`（按 Spec/UI/Iterate）与 `ChatFeed.tsx`（hooks + bubble/dock）
+- **下次优先：** D38 拆 `ChatFeed.tsx`（hooks + bubble/dock）；P2 `App.tsx` / `builtin_tools.py`
 
 ## Recent Sessions
+
+## 2026-08-01 会话（D38：design/generator Spec/UI/Iterate）
+
+- **落地：** `gen_spec.py` / `gen_ui.py` / `gen_iterate.py`；`generator.py` 编排壳 + re-export
+- **测：** `pytest tests/test_design_service.py tests/test_prototype_generator.py` 44 passed
+- **下次：** ChatFeed hooks + bubble/dock
 
 ## 2026-08-01 会话（D38 Phase 2：chat_runner plain/workflow）
 
@@ -448,6 +454,12 @@
 - （可选）配置 `HOMEBREW_TAP_GITHUB_TOKEN`
 
 ## Recent Sessions
+
+## 2026-08-01 会话（D38：design/generator Spec/UI/Iterate）
+
+- **落地：** `gen_spec.py` / `gen_ui.py` / `gen_iterate.py`；`generator.py` 编排壳 + re-export
+- **测：** `pytest tests/test_design_service.py tests/test_prototype_generator.py` 44 passed
+- **下次：** ChatFeed hooks + bubble/dock
 
 ## 2026-07-11 会话（v1.2.2 发版准备）
