+++ services/orchestrator/src/design/generator.py
@@ -76,62 +76,71 @@ def _normalize_preset_id(preset_id: str | None) -> str:
 
 logger = logging.getLogger(__name__)
 
+# D38 — stage modules (re-exported for service/tests)
+from src.design.gen_spec import (
+    _enhance_brief,
+    _extract_colors_from_design_md,
+    _extract_json_block,
+    _extract_json_data,
+    _fallback_spec,
+    _get_unified_tailwind_config,
+    _html_matches_brief_intent,
+    _parse_multi_screens,
+    _prompt_intent,
+    _spec_confirm_enabled,
+    _spec_to_design_md,
+    _detect_html_intent,
+    inject_unified_tailwind_config,
+)
+from src.design.gen_ui import (
+    _DESIGN_REVIEW_ENABLED,
+    _LLM_TIMEOUT_SEC,
+    _LLM_UI_TIMEOUT_SEC,
+    _TASTE_ANTI_PATTERNS,
+    _build_shared_layout_context,
+    _build_ui_compact_prompt,
+    _build_ui_correction_prompt,
+    _build_ui_generation_prompt,
+    _check_vision_ok,
+    _coerce_ui_html,
+    _design_review_and_improve,
+    _extract_css_tokens,
+    _extract_html_from_llm,
+    _fallback_login_html,
+    _fallback_ui_html,
+    _fetch_url_snapshot,
+    _first_hex,
+    _format_css_tokens_for_prompt,
+    _generate_ui_html,
+    _html_from_llm_response,
+    _html_has_visible_content,
+    _llm_complete,
+    _llm_complete_vision,
+    _looks_like_vision_error,
+    _shell_html,
+    _try_llm_complete_vision,
+)
+from src.design.gen_iterate import (
+    _html_essentially_same,
+    _infer_iterate_mode,
+    _merged_design_prompt,
+    _next_round_index,
+    _next_screen_id,
+    _record_screen_round,
+    _resolve_screen_html_path,
+    _screen_html_rel,
+    _screen_layout_x,
+    delete_screen,
+    iterate_session,
+)
+
+
 _generate_jobs: dict[str, threading.Thread] = {}
 _generate_lock = threading.Lock()
 _iterate_jobs: dict[str, threading.Thread] = {}
 _iterate_lock = threading.Lock()
 
 
-def _spec_confirm_enabled() -> bool:
-    """D40: Spec soft-confirm gate (default off). Set CLUTCH_DESIGN_SPEC_CONFIRM=1 to pause after Spec."""
-    return os.environ.get("CLUTCH_DESIGN_SPEC_CONFIRM", "0").strip().lower() in {
-        "1",
-        "true",
-        "yes",
-        "on",
-    }
-
-
-def _enhance_brief(prompt: str, device: str) -> str:
-    """Structure vague UI briefs before Spec/UI (local port of Stitch enhance-prompt)."""
-    text = (prompt or "").strip()
-    if not text:
-        return text
-    if "**PAGE STRUCTURE:**" in text.upper() or "**PLATFORM:**" in text.upper():
-        return text
-    platform = "Mobile, Mobile-first" if device == "app" else "Web, Desktop-first"
-    upgraded = text
-    for pat, repl in (
-        (r"(?i)\bmenu at the top\b", "navigation bar with logo and menu items"),
-        (r"(?i)\blist of items\b", "card grid layout"),
-        (r"(?i)\bpicture area\b", "hero section with full-width image"),
-    ):
-        upgraded = re.sub(pat, repl, upgraded)
-    return (
-        f"{upgraded}\n\n"
-        f"**PLATFORM:** {platform}\n\n"
-        "**PAGE STRUCTURE:** Infer numbered sections from the brief; preserve all user intent.\n"
-        "1. **Header / Nav**\n"
-        "2. **Primary content**\n"
-        "3. **Supporting sections / footer as needed**\n"
-    )
-
-
-_TASTE_ANTI_PATTERNS = (
-    "Taste / anti-patterns (must follow):\n"
-    "- No Inter font; no pure black #000000; no neon purple/blue glow CTAs.\n"
-    "- No generic 3 equal feature cards; prefer asymmetric or 2-column layouts.\n"
-    "- No AI copy clichés (Elevate, Seamless, Unleash, Next-Gen) or invented metrics.\n"
-    "- No emoji decoration; no 'Scroll to explore' filler.\n"
-)
-_LLM_TIMEOUT_SEC = 45.0
-_LLM_UI_TIMEOUT_SEC = 90.0
-_DESIGN_REVIEW_ENABLED = os.environ.get("CLUTCH_DESIGN_REVIEW", "").strip().lower() in {
-    "1",
-    "true",
-    "yes",
-    "on",
-}
 _DESIGN_ROW_Y = 56
 
 
@@ -338,1781 +347,6 @@ def _llm_result(
     text = str(result).strip()
     usage, estimated = usage_from_llm_result(result, prompt=prompt, response_text=text)
     return text, None, usage, estimated
-
-
-def _extract_json_data(text: str) -> Any:
-    raw = (text or "").strip()
-    if not raw:
-        return None
-    m = re.search(r"```(?:json)?\s*([\s\S]*?)```", raw, re.I)
-    if m:
-        raw = m.group(1).strip()
-    try:
-        return json.loads(raw)
-    except json.JSONDecodeError:
-        start_dict = raw.find("{")
-        end_dict = raw.rfind("}")
-        start_list = raw.find("[")
-        end_list = raw.rfind("]")
-        
-        # If both are present, pick the one that starts first
-        if start_dict >= 0 and (start_list < 0 or start_dict < start_list):
-            if end_dict > start_dict:
-                try:
-                    return json.loads(raw[start_dict : end_dict + 1])
-                except json.JSONDecodeError:
-                    pass
-        elif start_list >= 0:
-            if end_list > start_list:
-                try:
-                    return json.loads(raw[start_list : end_list + 1])
-                except json.JSONDecodeError:
-                    pass
-        raise
-
-
-def _extract_json_block(text: str) -> dict[str, Any]:
-    data = _extract_json_data(text)
-    if isinstance(data, dict):
-        return data
-    return {}
-
-
-def _parse_multi_screens(
-    prompt: str,
-    model_id: str,
-    router: Any,
-) -> list[dict[str, Any]]:
-    """
-    Parse a user prompt to see if they want multiple screens/pages.
-    Returns a list of dicts: [{"id": str, "name": str, "prompt": str, "layout_pattern": str}]
-    """
-    from src.models_config import is_model_available
-    default_screen = [{"id": "main", "name": "Interface", "prompt": prompt, "layout_pattern": None}]
-    if not is_model_available(router, model_id):
-        return default_screen
-
-    planning_prompt = (
-        "You are an AI UI/UX architect planning screens/pages for a design workflow.\n"
-        f"User prompt: {prompt}\n\n"
-        "Analyze the prompt. If the user wants multiple pages, screens, or features generated (for example, 'a homepage, a profile page, and a login page'), "
-        "plan a list of screens to generate.\n"
-        "If the user only wants a single screen or page, or if it is ambiguous, return a list with a single element representing the requested screen.\n\n"
-        "Return ONLY a valid JSON list of objects, representing each planned screen in order. "
-        "Each object must have the following keys:\n"
-        "- 'id': a short lowercase alphanumeric string (e.g. 'home', 'profile', 'login', 'dashboard', 'settings'). The first screen MUST have id 'main'.\n"
-        "- 'name': a user-friendly name in the prompt's language (e.g. '首页', '个人中心', '登录页面', 'Home', 'Profile', 'Login').\n"
-        "- 'prompt': a specific UI generation instruction for this screen, combining the global design system/topic with the screen's specific purpose.\n"
-        "- 'layout_pattern': one of 'landing', 'dashboard', 'crm', 'settings', 'analytics', 'ecommerce', 'chat', 'mobile_app', 'login', 'pricing', 'profile'.\n\n"
-        "Do NOT wrap in markdown code blocks like ```json or ```. Return ONLY the raw JSON string."
-    )
-
-    try:
-        raw, _, _, _ = _llm_complete(router, planning_prompt, model_id=model_id)
-        parsed = _extract_json_data(raw)
-        if isinstance(parsed, list) and len(parsed) > 0:
-            valid_screens = []
-            for i, item in enumerate(parsed):
-                if not isinstance(item, dict):
-                    continue
-                item_id = str(item.get("id") or "").strip().lower()
-                item_name = str(item.get("name") or "").strip()
-                item_prompt = str(item.get("prompt") or "").strip()
-                item_pattern = str(item.get("layout_pattern") or "").strip().lower()
-                
-                if not item_name or not item_prompt:
-                    continue
-                
-                if i == 0:
-                    sid = "main"
-                else:
-                    sid = re.sub(r"[^a-z0-9_-]", "", item_id) or f"screen-{i+1}"
-                    if sid == "main":
-                        sid = f"screen-{i+1}"
-                
-                if item_pattern not in {
-                    "landing", "dashboard", "crm", "settings", "analytics",
-                    "ecommerce", "chat", "mobile_app", "login", "pricing", "profile"
-                }:
-                    item_pattern = None
-
-                valid_screens.append({
-                    "id": sid,
-                    "name": item_name,
-                    "prompt": item_prompt,
-                    "layout_pattern": item_pattern
-                })
-            if valid_screens:
-                return valid_screens
-    except Exception as exc:
-        logger.warning("Failed to parse multi-screen plan, falling back to single: %s", exc)
-
-    return default_screen
-
-
-def _fallback_spec(prompt: str) -> dict[str, Any]:
-    return {
-        "name": "Design Spec",
-        "rationale": f"Fallback offline spec for prompt: {prompt[:80]}",
-        "colors": {
-            "primary": ["#2563eb", "#1d4ed8"],
-            "secondary": ["#64748b", "#475569"],
-            "neutral": ["#ffffff", "#f8fafc", "#0f172a"],
-        },
-        "typography": {
-            "fontFamily": "Inter, system-ui, sans-serif",
-            "samples": [
-                {"label": "Heading", "size": "1.875rem", "weight": "700"},
-                {"label": "Body", "size": "1rem", "weight": "400"},
-            ],
-        },
-        "components": ["Header", "Card Grid", "Footer"],
-    }
-
-
-def _prompt_intent(prompt: str) -> str:
-    """Coarse UI intent from the user brief — drives fallback HTML, not LLM."""
-    p = (prompt or "").strip().lower()
-    if any(
-        k in p
-        for k in (
-            "登录",
-            "登陆",
-            "注册",
-            "signin",
-            "sign-in",
-            "sign in",
-            "log in",
-            "login",
-            "sign up",
-            "signup",
-            "auth",
-        )
-    ):
-        return "login"
-    if any(
-        k in p
-        for k in (
-            "购物",
-            "商城",
-            "电商",
-            "商品",
-            "shop",
-            "store",
-            "ecommerce",
-            "e-commerce",
-            "product",
-            "cart",
-            "marketplace",
-        )
-    ):
-        return "shop"
-    if any(k in p for k in ("仪表", "dashboard", "后台", "admin", "analytics", "控制台")):
-        return "dashboard"
-    if any(
-        k in p
-        for k in (
-            "音乐",
-            "播放器",
-            "歌词",
-            "歌单",
-            "切歌",
-            "music",
-            "player",
-            "playlist",
-            "lyrics",
-            "spotify",
-            "song",
-            "album",
-        )
-    ):
-        return "music"
-    if any(k in p for k in ("落地", "landing", "官网", "首页", "home page", "marketing", "hero")):
-        return "landing"
-    return "generic"
-
-
-def _detect_html_intent(html: str) -> str | None:
-    l = html.lower()
-    if "type=\"password\"" in l or "placeholder=\"••••••••\"" in l or "sign in" in l or "log in" in l:
-        return "login"
-    if "cart" in l or "product" in l or "price" in l or "shop" in l:
-        return "shop"
-    if "chart" in l or "dashboard" in l or "graph" in l or "metrics" in l:
-        return "dashboard"
-    if "playlist" in l or "lyrics" in l or "track" in l or "music" in l:
-        return "music"
-    return None
-
-
-def _html_matches_brief_intent(prompt: str, html: str) -> bool:
-    expected = _prompt_intent(prompt)
-    detected = _detect_html_intent(html)
-    if expected == "landing" and detected == "login":
-        return False
-    if expected in {"landing", "generic"}:
-        return True
-    return detected == expected or detected is None
-
-
-def _screen_html_rel(screen: dict[str, Any]) -> str:
-    sid = str(screen.get("id") or "main")
-    return str(screen.get("html_path") or f"screens/{sid}.html")
-
-
-def _resolve_screen_html_path(session_dir: Path, screen: dict[str, Any]) -> Path:
-    return session_dir / _screen_html_rel(screen)
-
-
-def _next_round_index(manifest: dict[str, Any], screen_id: str) -> int:
-    history = list(manifest.get("round_history") or [])
-    indices = [
-        int(r.get("round_index", 0))
-        for r in history
-        if str(r.get("screen_id") or "") == screen_id
-    ]
-    return (max(indices) + 1) if indices else 0
-
-
-def _record_screen_round(
-    session_dir: Path,
-    manifest: dict[str, Any],
-    *,
-    screen_id: str,
-    html: str,
-    prompt: str,
-    reasoning_content: str | None,
-    process_log_slice: list[dict[str, Any]],
-) -> dict[str, Any]:
-    """Write versioned HTML and append round metadata to manifest."""
-    round_index = _next_round_index(manifest, screen_id)
-    rel = f"screens/{screen_id}_r{round_index}.html"
-    (session_dir / "screens").mkdir(exist_ok=True)
-    (session_dir / rel).write_text(html, encoding="utf-8")
-    entry: dict[str, Any] = {
-        "round_index": round_index,
-        "screen_id": screen_id,
-        "html_path": rel,
-        "prompt": prompt,
-        "reasoning_content": reasoning_content,
-        "process_log": process_log_slice,
-        "at": now_iso(),
-    }
-    history = list(manifest.get("round_history") or [])
-    history.append(entry)
-    manifest["round_history"] = history
-    for screen in manifest.get("screens") or []:
-        if str(screen.get("id")) == screen_id:
-            screen["html_path"] = rel
-            screen["active_round_index"] = round_index
-    return entry
-
-
-def _first_hex(colors: dict[str, Any] | None, key: str, fallback: str) -> str:
-    if not colors:
-        return fallback
-    values = colors.get(key)
-    if isinstance(values, list) and values:
-        raw = str(values[0]).strip()
-        if re.match(r"^#([0-9a-fA-F]{3}|[0-9a-fA-F]{6})$", raw):
-            return raw
-    if isinstance(values, str) and re.match(r"^#([0-9a-fA-F]{3}|[0-9a-fA-F]{6})$", values.strip()):
-        return values.strip()
-    return fallback
-
-
-def _shell_html(title: str, body: str, *, device: str = "web") -> str:
-    is_app = (device or "web").strip().lower() == "app"
-    canvas = (
-        "width:390px;min-height:844px;margin:0 auto;"
-        if is_app
-        else "width:1920px;min-height:1080px;margin:0 auto;"
-    )
-    return f"""<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="utf-8"/>
-<meta name="viewport" content="width={'390' if is_app else '1920'}, initial-scale=1"/>
-<title>{esc(title)}</title>
-<script src="https://cdn.tailwindcss.com"></script>
-<style>
-  html, body {{ margin:0; background:#f8fafc; font-family: system-ui, -apple-system, sans-serif; }}
-  .clutch-canvas {{ {canvas} background:#fff; }}
-</style>
-</head>
-<body>
-<div class="clutch-canvas">
-{body}
-</div>
-</body>
-</html>
-"""
-
-
-def _html_has_visible_content(html: str) -> bool:
-    """True when the document body has real UI — not an empty shell after wrap."""
-    if not (html or "").strip():
-        return False
-    lower = html.lower()
-    if "<html" in lower and "<body" not in lower:
-        return False
-    if "<style" in lower and "</style>" not in lower:
-        return False
-    if "<script" in lower and "</script>" not in lower:
-        return False
-    m = re.search(
-        r'class=["\']clutch-canvas["\'][^>]*>([\s\S]*?)</div>\s*</body>',
-        html,
-        re.I,
-    )
-    if not m:
-        m = re.search(r"<body[^>]*>([\s\S]*?)</body>", html, re.I)
-    chunk = m.group(1) if m else html
-    chunk = re.sub(r"<script[\s\S]*?</script>", "", chunk, flags=re.I)
-    chunk = re.sub(r"<style[\s\S]*?</style>", "", chunk, flags=re.I)
-    chunk = re.sub(r"<!--[\s\S]*?-->", "", chunk)
-    compact = re.sub(r">\s+<", "><", chunk.strip())
-    if not compact:
-        return False
-    if re.fullmatch(r"(?:<div\b[^>]*>\s*</div>\s*)+", compact, flags=re.I):
-        return False
-    text = re.sub(r"\s+", "", re.sub(r"<[^>]+>", "", chunk))
-    if len(text) >= 2:
-        return True
-    return bool(
-        re.search(
-            r"<(?:div|section|main|header|nav|footer|article|aside|h[1-6]|p|button|a|"
-            r"img|ul|ol|li|form|input|table|span|svg)\b[^>]*>",
-            chunk,
-            re.I,
-        )
-    )
-
-
-def _fallback_login_html(prompt: str, spec: dict[str, Any]) -> str:
-    return "<div>Login stub</div>"
-
-
-def _spec_to_design_md(spec: dict[str, Any]) -> str:
-    lines = [
-        f"# DESIGN SPECIFICATION: {spec.get('name') or 'Design System'}",
-        "",
-        "## Rationale",
-        str(spec.get("rationale") or "Polished visual structure."),
-        "",
-        "## Colors",
-    ]
-    colors = spec.get("colors") or {}
-    for k, v in colors.items():
-        if isinstance(v, list) and v:
-            lines.append(f"- **{k}**: {', '.join(v)}")
-        elif v:
-            lines.append(f"- **{k}**: {v}")
-    lines += [
-        "",
-        "## Typography",
-        f"- **Font Family**: {spec.get('typography', {}).get('fontFamily', 'Inter')}",
-        "",
-        "## Components",
-    ]
-    for comp in spec.get("components") or []:
-        lines.append(f"- {comp}")
-    lines += [
-        "",
-        "## Accessibility",
-        str(spec.get("accessibility") or "WCAG AA contrast; focus rings; semantic headings."),
-        "",
-    ]
-    return "\n".join(lines)
-
-
-def _extract_colors_from_design_md(design_md: str) -> dict[str, str]:
-    colors = {}
-    if not design_md:
-        return colors
-    in_colors = False
-    for line in design_md.splitlines():
-        line_stripped = line.strip()
-        if not line_stripped:
-            continue
-        if line_stripped.startswith("colors:"):
-            in_colors = True
-            continue
-        if in_colors:
-            if line_stripped.startswith("typography:") or line_stripped.startswith("#") or (not line.startswith(" ") and not line.startswith("-") and ":" not in line_stripped):
-                in_colors = False
-                continue
-            match = re.search(r"^\s*-\s*([^:]+):\s*['\"]?(#[0-9a-fA-F]{3,8}|[a-zA-Z0-9]+)['\"]?", line)
-            if not match:
-                match = re.search(r"^\s*([^:]+):\s*['\"]?(#[0-9a-fA-F]{3,8}|[a-zA-Z0-9]+)['\"]?", line)
-            if match:
-                k = match.group(1).strip().strip("-").strip()
-                v = match.group(2).strip()
-                colors[k] = v
-    return colors
-
-
-def _get_unified_tailwind_config(spec: dict[str, Any] | None, design_md: str) -> dict[str, Any]:
-    colors = {}
-    if spec and isinstance(spec, dict):
-        raw_colors = spec.get("colors") or {}
-        for k, v in raw_colors.items():
-            if isinstance(v, list):
-                if len(v) > 0:
-                    colors[k] = v[0]
-                    if len(v) > 1:
-                        colors[f"{k}-light"] = v[1]
-                        colors[f"{k}-hover"] = v[1]
-                    if len(v) > 2:
-                        colors[f"{k}-dark"] = v[2]
-                        colors[f"{k}-active"] = v[2]
-            elif isinstance(v, str):
-                colors[k] = v
-                
-    md_colors = _extract_colors_from_design_md(design_md)
-    for k, v in md_colors.items():
-        if k not in colors:
-            colors[k] = v
-
-    theme_config = {
-        "colors": colors,
-        "borderRadius": {
-            "DEFAULT": "0.125rem",
-            "sm": "0.25rem",
-            "md": "0.375rem",
-            "lg": "0.5rem",
-            "xl": "0.75rem",
-            "2xl": "1rem",
-            "3xl": "1.5rem",
-            "full": "9999px"
-        },
-        "spacing": {
-            "unit": "4px",
-            "xs": "4px",
-            "sm": "8px",
-            "md": "16px",
-            "lg": "24px",
-            "xl": "32px",
-            "2xl": "48px",
-            "3xl": "64px",
-            "container-max": "1440px",
-            "sidebar-width": "240px",
-            "gutter": "16px"
-        }
-    }
-    
-    if spec and isinstance(spec, dict):
-        font_family = spec.get("typography", {}).get("fontFamily")
-        if font_family:
-            theme_config["fontFamily"] = {
-                "sans": [font_family, "Inter", "system-ui", "-apple-system", "sans-serif"],
-                "body": [font_family, "Inter", "system-ui", "-apple-system", "sans-serif"]
-            }
-    return theme_config
-
-
-def inject_unified_tailwind_config(html: str, spec: dict[str, Any] | None, design_md: str) -> str:
-    if not html:
-        return html
-    theme_config = _get_unified_tailwind_config(spec, design_md)
-    theme_config_json = json.dumps(theme_config, indent=6, ensure_ascii=False)
-    
-    script_content = f"""<script id="tailwind-config">
-      tailwind.config = {{
-        darkMode: "class",
-        theme: {{
-          extend: {theme_config_json}
-        }}
-      }};
-    </script>"""
-    
-    pattern = re.compile(
-        r'<script[^>]*id=["\']tailwind-config["\'][^>]*>[\s\S]*?</script>|<script[^>]*>\s*tailwind\.config\s*=\s*\{[\s\S]*?\}\s*;?\s*</script>',
-        re.I
-    )
-    if pattern.search(html):
-        html = pattern.sub(script_content, html)
-    else:
-        cdn_pattern = re.compile(r'(<script[^>]*src=["\'][^"\']*tailwindcss\.com[^"\']*["\'][^>]*>\s*</script>)', re.I)
-        if cdn_pattern.search(html):
-            html = cdn_pattern.sub(r'\1\n' + script_content, html)
-        elif "</head>" in html:
-            html = html.replace("</head>", f"{script_content}\n</head>")
-        else:
-            html = f"{script_content}\n{html}"
-    return html
-
-
-
-def _fallback_ui_html(prompt: str, spec: dict[str, Any], *, device: str = "web") -> str:
-    """Offline-only stub when no model/API key is configured."""
-    primary = (spec.get("colors") or {}).get("primary") or ["#2563eb"]
-    accent = primary[0] if isinstance(primary, list) else str(primary)
-    title = (prompt.strip() or str(spec.get("name") or "Welcome")).split("\n")[0][:48]
-    intent = _prompt_intent(prompt)
-    is_app = (device or "web").strip().lower() == "app"
-    shell = "px-4" if is_app else "px-16"
-    cols = "grid-cols-2" if is_app else "grid-cols-4"
-    cards = "".join(
-        f"""
-      <div class="rounded-2xl border border-slate-200 bg-white overflow-hidden shadow-sm">
-        <div class="{'h-28' if is_app else 'h-44'}" style="background:{accent};opacity:0.85"></div>
-        <div class="p-3 space-y-1">
-          <p class="text-sm font-semibold">Product {i}</p>
-          <p class="text-xs text-slate-500">¥{(i * 129)}</p>
-          <button style="background:{accent}" class="mt-2 w-full rounded-lg text-white text-xs font-semibold py-1.5">Add to cart</button>
-        </div>
-      </div>"""
-        for i in range(1, 5)
-    )
-
-    if intent == "login":
-        body = f"""
-<div class="min-h-full flex items-center justify-center p-6 {shell}">
-  <div class="w-full {'max-w-sm' if is_app else 'max-w-md'} space-y-5 rounded-2xl border border-slate-200 bg-white p-6 shadow-sm">
-    <div>
-      <h1 class="text-2xl font-bold">Welcome back</h1>
-      <p class="text-sm text-slate-500 mt-1">{esc(title)}</p>
-    </div>
-    <label class="block text-xs font-semibold text-slate-600">Email
-      <input class="mt-1 w-full rounded-xl border border-slate-200 px-3 py-2.5 text-sm" placeholder="you@company.com"/>
-    </label>
-    <label class="block text-xs font-semibold text-slate-600">Password
-      <input type="password" class="mt-1 w-full rounded-xl border border-slate-200 px-3 py-2.5 text-sm" placeholder="••••••••"/>
-    </label>
-    <button style="background:{accent}" class="w-full rounded-xl text-white font-semibold py-2.5 text-sm">Log in</button>
-  </div>
-</div>
-"""
-    elif intent == "shop":
-        body = f"""
-<div class="min-h-full">
-  <header class="py-4 border-b border-slate-100 flex items-center justify-between {shell}">
-    <p class="font-bold text-sm">{esc(str(spec.get("name") or "Store"))}</p>
-    <span class="text-xs text-slate-500">Cart (0)</span>
-  </header>
-  <main class="py-8 {shell}">
-    <h1 class="{'text-xl' if is_app else 'text-3xl'} font-bold mb-1">Featured</h1>
-    <p class="text-sm text-slate-500 mb-6">Shop the latest picks</p>
-    <div class="grid {cols} gap-4">{cards}
-    </div>
-  </main>
-</div>
-"""
-    elif intent == "dashboard":
-        body = f"""
-<div class="min-h-full flex {'flex-col' if is_app else ''}">
-  <aside class="{'w-full border-b' if is_app else 'w-64 border-r min-h-[1080px]'} bg-slate-900 text-white p-5 space-y-3">
-    <p class="font-bold text-sm">{esc(title)}</p>
-    <p class="text-xs text-slate-400">Overview</p>
-    <p class="text-xs text-slate-400">Orders</p>
-    <p class="text-xs text-slate-400">Settings</p>
-  </aside>
-  <main class="flex-1 p-6 space-y-4">
-    <h1 class="text-2xl font-bold">Dashboard</h1>
-    <div class="grid grid-cols-2 {'md:grid-cols-3' if not is_app else ''} gap-4">
-      <div class="rounded-xl border border-slate-200 bg-white p-4"><p class="text-xs text-slate-500">Revenue</p><p class="text-lg font-bold">¥128k</p></div>
-      <div class="rounded-xl border border-slate-200 bg-white p-4"><p class="text-xs text-slate-500">Orders</p><p class="text-lg font-bold">1,284</p></div>
-      <div class="rounded-xl border border-slate-200 bg-white p-4"><p class="text-xs text-slate-500">Users</p><p class="text-lg font-bold">8.2k</p></div>
-    </div>
-    <div class="h-56 rounded-xl border border-slate-200 bg-white" style="background:linear-gradient(135deg,{accent}22,transparent)"></div>
-  </main>
-</div>
-"""
-    elif intent == "music":
-        tracks = "".join(
-            f"""
-      <button type="button" class="w-full flex items-center gap-3 rounded-xl px-3 py-2.5 text-left hover:bg-white/10 {'bg-white/15' if i == 1 else ''}">
-        <span class="flex h-10 w-10 items-center justify-center rounded-lg text-xs font-bold text-white" style="background:{accent}">{i}</span>
-        <span class="min-w-0 flex-1">
-          <span class="block truncate text-sm font-semibold text-white">Track {i} · Night Drive</span>
-          <span class="block truncate text-xs text-white/50">Artist {i}</span>
-        </span>
-        <span class="text-[10px] text-white/40">3:2{i}</span>
-      </button>"""
-            for i in range(1, 6)
-        )
-        if is_app:
-            body = f"""
-<div class="min-h-full text-white" style="background:linear-gradient(180deg,#0f172a,#020617)">
-  <header class="px-4 pt-6 pb-3 flex items-center justify-between">
-    <p class="text-sm font-bold">{esc(title)}</p>
-    <span class="text-xs text-white/50">Library</span>
-  </header>
-  <main class="px-4 space-y-4 pb-28">
-    <div class="rounded-2xl p-4" style="background:linear-gradient(135deg,{accent},#111827)">
-      <p class="text-xs text-white/70">Now playing</p>
-      <p class="mt-1 text-lg font-bold">Midnight Pulse</p>
-      <p class="text-xs text-white/60">SonicFlow · Album</p>
-    </div>
-    <section>
-      <p class="mb-2 text-xs font-semibold uppercase tracking-wide text-white/50">Playlist</p>
-      <div class="space-y-1">{tracks}</div>
-    </section>
-    <section class="rounded-2xl border border-white/10 bg-white/5 p-4 backdrop-blur">
-      <p class="mb-2 text-xs font-semibold text-white/60">Lyrics</p>
-      <p class="text-sm leading-relaxed text-white/90">City lights blur past the glass…</p>
-      <p class="mt-2 text-sm leading-relaxed text-white/50">Hold the night a little longer…</p>
-      <p class="mt-2 text-sm leading-relaxed text-white/35">Pulse under neon rain…</p>
-    </section>
-  </main>
-  <footer class="fixed bottom-0 left-0 right-0 border-t border-white/10 bg-black/80 px-4 py-3 backdrop-blur">
-    <div class="flex items-center justify-between gap-3">
-      <button type="button" class="rounded-full bg-white/10 px-3 py-2 text-xs">Prev</button>
-      <button type="button" class="rounded-full px-5 py-2 text-xs font-bold text-white" style="background:{accent}">Play</button>
-      <button type="button" class="rounded-full bg-white/10 px-3 py-2 text-xs">Next</button>
-    </div>
-  </footer>
-</div>
-"""
-        else:
-            body = f"""
-<div class="min-h-full flex text-white" style="background:#020617">
-  <aside class="w-56 border-r border-white/10 p-5 space-y-3">
-    <p class="font-bold">{esc(str(spec.get("name") or "SonicFlow"))}</p>
-    <p class="text-xs text-white/50">Home</p>
-    <p class="text-xs text-white/50">Search</p>
-    <p class="text-xs text-white/80">Your Library</p>
-  </aside>
-  <main class="flex-1 flex flex-col min-h-[1080px]">
-    <div class="flex-1 grid grid-cols-2 gap-6 p-8">
-      <section class="space-y-4">
-        <div class="rounded-3xl p-8" style="background:linear-gradient(135deg,{accent},#1e1b4b)">
-          <p class="text-sm text-white/70">Featured</p>
-          <h1 class="mt-2 text-4xl font-bold">Midnight Pulse</h1>
-          <p class="mt-2 text-white/60">A dark immersive player with playlist + lyrics.</p>
-        </div>
-        <div>
-          <p class="mb-3 text-sm font-semibold text-white/70">Playlist</p>
-          <div class="space-y-1">{tracks}</div>
-        </div>
-      </section>
-      <section class="rounded-3xl border border-white/10 bg-white/5 p-6 backdrop-blur flex flex-col">
-        <p class="text-sm font-semibold text-white/70">Lyrics</p>
-        <div class="mt-6 flex-1 space-y-4 text-lg leading-relaxed">
-          <p class="text-white">City lights blur past the glass…</p>
-          <p class="text-white/70">Hold the night a little longer…</p>
-          <p class="text-white/45">Pulse under neon rain…</p>
-          <p class="text-white/30">Switch tracks — keep the vibe.</p>
-        </div>
-      </section>
-    </div>
-    <footer class="border-t border-white/10 px-8 py-4 flex items-center justify-between bg-black/50">
-      <div>
-        <p class="text-sm font-semibold">Midnight Pulse</p>
-        <p class="text-xs text-white/50">Now playing</p>
-      </div>
-      <div class="flex items-center gap-3">
-        <button type="button" class="rounded-full bg-white/10 px-4 py-2 text-sm">Prev</button>
-        <button type="button" class="rounded-full px-6 py-2 text-sm font-bold" style="background:{accent}">Play</button>
-        <button type="button" class="rounded-full bg-white/10 px-4 py-2 text-sm">Next</button>
-      </div>
-    </footer>
-  </main>
-</div>
-"""
-    else:
-        body = f"""
-<div class="min-h-full">
-  <header class="py-5 flex items-center justify-between {shell}">
-    <p class="font-bold text-sm">{esc(str(spec.get("name") or "Brand"))}</p>
-    <button style="background:{accent}" class="rounded-lg text-white text-xs font-semibold px-4 py-2">Get started</button>
-  </header>
-  <main class="py-16 {shell} text-center space-y-5">
-    <h1 class="{'text-3xl' if is_app else 'text-5xl'} font-bold tracking-tight">{esc(title)}</h1>
-    <p class="text-slate-500 {'text-sm' if is_app else 'text-lg'} max-w-2xl mx-auto">A polished interface draft for your brief. Iterate from the canvas to refine layout and copy.</p>
-    <div class="flex justify-center gap-3">
-      <button style="background:{accent}" class="rounded-xl text-white font-semibold px-6 py-3 text-sm">Primary action</button>
-      <button class="rounded-xl border border-slate-200 px-6 py-3 text-sm font-medium">Learn more</button>
-    </div>
-    <div class="mt-10 grid grid-cols-1 {'md:grid-cols-3' if not is_app else ''} gap-4 text-left">
-      <div class="rounded-2xl border border-slate-200 bg-white p-5"><p class="font-semibold text-sm mb-1">Feature A</p><p class="text-xs text-slate-500">Describe value here.</p></div>
-      <div class="rounded-2xl border border-slate-200 bg-white p-5"><p class="font-semibold text-sm mb-1">Feature B</p><p class="text-xs text-slate-500">Describe value here.</p></div>
-      <div class="rounded-2xl border border-slate-200 bg-white p-5"><p class="font-semibold text-sm mb-1">Feature C</p><p class="text-xs text-slate-500">Describe value here.</p></div>
-    </div>
-  </main>
-</div>
-"""
-    return _shell_html(title, body, device=device)
-
-
-def _coerce_ui_html(
-    raw: str,
-    *,
-    title: str,
-    prompt: str,
-    spec: dict[str, Any],
-    device: str,
-    fallback_html: str | None = None,
-    allow_template_fallback: bool = True,
-) -> str:
-    """Wrap fragment if needed; optional offline template when LLM output is blank."""
-    html = (raw or "").strip()
-    if html and "<html" not in html.lower():
-        html = _shell_html(title, html, device=device)
-    if html and "charset" not in html.lower()[:500]:
-        if "<head>" in html.lower():
-            html = html.replace("<head>", '<head>\n<meta charset="utf-8"/>', 1)
-        elif "<head " in html.lower():
-            import re as _re
-            html = _re.sub(r"(<head\s[^>]*>)", r'\1\n<meta charset="utf-8"/>', html, count=1, flags=_re.I)
-    if _html_has_visible_content(html):
-        return html
-    if fallback_html is not None and _html_has_visible_content(fallback_html):
-        return fallback_html
-    if allow_template_fallback:
-        return _fallback_ui_html(prompt, spec, device=device)
-    return ""
-
-
-def _llm_complete(
-    router: Any, prompt: str, *, model_id: str, timeout_sec: float = _LLM_TIMEOUT_SEC
-) -> tuple[str, str | None, dict[str, int], bool]:
-    """Run router.complete with a hard timeout; returns content, reasoning, usage, estimated."""
-    from concurrent.futures import ThreadPoolExecutor, TimeoutError as FuturesTimeout
-
-    with ThreadPoolExecutor(max_workers=1) as pool:
-        future = pool.submit(router.complete, prompt, model_id=model_id)
-        try:
-            return _llm_result(future.result(timeout=timeout_sec), prompt=prompt)
-        except FuturesTimeout as exc:
-            future.cancel()
-            raise DesignError(f"Model timed out after {int(timeout_sec)}s") from exc
-
-
-def _check_vision_ok(router: Any, model_id: str, image_data_url: str | None) -> bool:
-    """Return True only if the model provider is known to support multimodal vision."""
-    if not image_data_url:
-        return False
-    try:
-        from src.adapters.ollama_adapter import model_supports_vision
-        spec, _ = router.resolve_for_model(model_id)
-        return model_supports_vision(spec)
-    except Exception:
-        return False
-
-
-def _looks_like_vision_error(text: str) -> bool:
-    """Return True if LLM output looks like a vision-capability error instead of real content."""
-    from src.chat_content import looks_like_vision_error
-
-    return looks_like_vision_error(text)
-
-
-def _llm_complete_vision(
-    router: Any,
-    prompt: str,
-    *,
-    model_id: str,
-    image_data_url: str | None = None,
-    timeout_sec: float = _LLM_TIMEOUT_SEC,
-    image_analysis_text: str = "",
-) -> tuple[str, str | None, dict[str, int], bool]:
-    """Complete with optional vision image via router.chat multimodal content.
-
-    Always tries sending the image first when present; on soft/API vision failure,
-    retries with local image analysis prefix (OCR/palette).
-    """
-    from concurrent.futures import ThreadPoolExecutor, TimeoutError as FuturesTimeout
-    from src.chat_content import looks_like_vision_api_error, user_message_content_for_llm
-
-    if image_data_url:
-        content = user_message_content_for_llm(
-            f"[image: {image_data_url}]\n{prompt}",
-            vision_enabled=True,
-        )
-        messages = [{"role": "user", "content": content}]
-
-        def _call() -> object:
-            return router.chat(messages, model_id=model_id)
-
-        try:
-            with ThreadPoolExecutor(max_workers=1) as pool:
-                future = pool.submit(_call)
-                try:
-                    result = _llm_result(future.result(timeout=timeout_sec), prompt=prompt)
-                except FuturesTimeout as exc:
-                    future.cancel()
-                    raise DesignError(f"Model timed out after {int(timeout_sec)}s") from exc
-
-            text, reasoning, usage, estimated = result
-            if not _looks_like_vision_error(text):
-                return result
-            logger.warning(
-                "vision call returned error-like response (model=%s) — falling back to local analysis",
-                model_id,
-            )
-        except DesignError:
-            raise
-        except Exception as exc:
-            if not looks_like_vision_api_error(exc):
-                raise
-            logger.warning(
-                "vision API error (model=%s) — falling back to local analysis: %s",
-                model_id,
-                exc,
-            )
-
-    prefix = ""
-    if image_data_url:
-        if image_analysis_text:
-            prefix = image_analysis_text + "\n\n"
-        else:
-            try:
-                from src.design.image_analysis import image_analysis_prompt_fragment
-                analysis = image_analysis_prompt_fragment(image_data_url)
-                prefix = (analysis + "\n\n") if analysis else (
-                    "A reference screenshot was attached; the model could not process vision input. "
-                    "Use the product brief to infer colors and layout.\n"
-                )
-            except Exception:
-                prefix = (
-                    "A reference screenshot was attached but the model could not process it. "
-                    "Infer a polished UI that matches the product brief alone.\n"
-                )
-    return _llm_complete(router, prefix + prompt, model_id=model_id, timeout_sec=timeout_sec)
-
-
-def _try_llm_complete_vision(
-    router: Any,
-    prompt: str,
-    *,
-    model_id: str,
-    image_data_url: str | None = None,
-    timeout_sec: float = _LLM_UI_TIMEOUT_SEC,
-) -> tuple[str, str | None, dict[str, int], bool, str | None]:
-    """Best-effort LLM call for UI HTML; never aborts pipeline."""
-    try:
-        text, reasoning, usage, estimated = _llm_complete_vision(
-            router,
-            prompt,
-            model_id=model_id,
-            image_data_url=image_data_url,
-            timeout_sec=timeout_sec,
-        )
-        return text, reasoning, usage, estimated, None
-    except DesignError as exc:
-        logger.warning("design ui llm call failed model=%s err=%s", model_id, exc)
-        return "", None, empty_token_usage(), False, str(exc)
-    except Exception as exc:
-        logger.warning("design ui llm call error model=%s err=%s", model_id, exc)
-        return "", None, empty_token_usage(), False, str(exc)
-
-
-def _extract_html_from_llm(text: str) -> str:
-    raw = (text or "").strip()
-    if not raw:
-        return ""
-
-    stripped = re.sub(r"^```[\w#\-]*\s*\n?", "", raw, flags=re.I, count=1)
-    stripped = re.sub(r"\n?```\s*$", "", stripped, flags=re.I)
-
-    for candidate in (stripped, raw):
-        doc = re.search(r"(<!DOCTYPE[\s\S]*?</html>)", candidate, re.I)
-        if doc:
-            return doc.group(1).strip()
-        html_open = re.search(r"(<html[\s\S]*?</html>)", candidate, re.I)
-        if html_open:
-            return html_open.group(1).strip()
-
-    fence = re.search(r"```(?:html)?\s*([\s\S]*?)```", raw, re.I)
-    if fence:
-        body = fence.group(1).strip()
-        if len(body) > 20 or re.search(r"<[a-z!]", body, re.I):
-            return body
-
-    return stripped
-
-
-def _html_from_llm_response(text: str, reasoning: str | None = None) -> str:
-    tag_re = re.compile(r"<(?:html|body|div|main|section|header|nav|form|button|h[1-6]|p)\b", re.I)
-    for chunk in (text, reasoning or ""):
-        if not (chunk or "").strip():
-            continue
-        extracted = _extract_html_from_llm(chunk)
-        if extracted and tag_re.search(extracted):
-            return extracted
-    return ""
-
-
-def _extract_css_tokens(html: str, base_url: str = "") -> dict[str, Any]:
-    """Extract real design tokens from raw HTML: colors, fonts, CSS vars, Tailwind classes."""
-    from collections import Counter
-    from html import unescape
-    from urllib.parse import urljoin, urlparse
-
-    tokens: dict[str, Any] = {}
-
-    style_blocks = re.findall(r"<style[^>]*>(.*?)</style>", html, re.I | re.S)
-    all_css = " ".join(style_blocks)
-
-    css_vars: dict[str, str] = {}
-    for k, v in re.findall(r"--([\.\w-]+)\s*:\s*([^;}\n]{1,80})", all_css):
-        css_vars[k.strip()] = v.strip().rstrip(",")
-    if css_vars:
-        tokens["css_vars"] = css_vars
-
-    css_var_refs: set[str] = set()
-    for ref in re.findall(r"var\(--([a-zA-Z][\w-]*)\)", html):
-        css_var_refs.add(ref)
-    if css_var_refs:
-        tokens["css_var_references"] = sorted(css_var_refs)
-
-    font_families: list[str] = []
-    for ff in re.findall(r"font-family\s*:\s*([^;}\n]+)", all_css):
-        cleaned = re.sub(r"['\"/]", "", ff.split(",")[0]).strip()
-        if cleaned and cleaned not in font_families and len(cleaned) < 60:
-            font_families.append(cleaned)
-    for family in re.findall(r"family=([A-Za-z+]+)", html):
-        name = family.replace("+", " ")
-        if name not in font_families:
-            font_families.append(name)
-    if font_families:
-        tokens["font_families"] = font_families[:6]
-
-    hex_counts: Counter = Counter()
-    for h in re.findall(r"#([0-9a-fA-F]{6}|[0-9a-fA-F]{3})\b", html):
-        norm = h.upper()
-        if len(norm) == 3:
-            norm = norm[0]*2 + norm[1]*2 + norm[2]*2
-        hex_counts["#" + norm] += 1
-    if hex_counts:
-        tokens["hex_colors"] = [h for h, _ in hex_counts.most_common(16)]
-        tokens["hex_colors_with_count"] = [[h, cnt] for h, cnt in hex_counts.most_common(16)]
-
-    tw_palette: Counter = Counter()
-    for cls in re.findall(r'(?:bg|text|border|ring|fill|stroke|from|to|via)-([a-z]+-[0-9]{2,3})\b', html):
-        tw_palette[cls] += 1
-    if tw_palette:
-        top_tw = [cls for cls, _ in tw_palette.most_common(12)]
-        tokens["tailwind_color_classes"] = top_tw
-        tw_theme_color = tw_palette.most_common(1)[0][0].split("-")[0] if tw_palette else ""
-        if tw_theme_color:
-            tokens["tailwind_theme_color"] = tw_theme_color
-
-    theme_m = re.search(
-        r'<meta[^>]+(name|property)=["\'](theme-color|msapplication-TileColor)["\'][^>]+content=["\']([^"\'>]+)',
-        html, re.I
-    ) or re.search(
-        r'<meta[^>]+content=["\']([^"\'>]+)["\'][^>]+(name|property)=["\'](theme-color|msapplication-TileColor)',
-        html, re.I
-    )
-    if theme_m:
-        tokens["theme_color"] = theme_m.group(3).strip()
-
-    og_m = re.search(r'og:image["\'][^>]*content=["\']([^"\'>]+)', html, re.I)
-    if not og_m:
-        og_m = re.search(r'content=["\']([^"\'>]+)["\'][^>]*og:image', html, re.I)
-    if og_m:
-        og_img = og_m.group(1).strip()
-        if og_img.startswith("http"):
-            tokens["og_image"] = og_img
-        elif base_url and og_img:
-            tokens["og_image"] = urljoin(base_url, og_img)
-
-    dark_signals = (
-        "class=\"dark\"" in html
-        or "data-theme=\"dark\"" in html
-        or re.search(r"@media\s*\(\s*prefers-color-scheme\s*:\s*dark\s*\)", all_css) is not None
-    )
-    dark_tw = len(re.findall(r'\bdark:', html)) > 5
-    # If theme-color meta explicitly says a light colour, the page is light regardless of dark: prefixes
-    tc = tokens.get("theme_color", "")
-    tc_is_light = False
-    if tc and re.match(r'^#([0-9a-fA-F]{6}|[0-9a-fA-F]{3})$', tc):
-        h = tc.lstrip("#")
-        if len(h) == 3:
-            h = h[0]*2 + h[1]*2 + h[2]*2
-        r, g, b = int(h[0:2], 16), int(h[2:4], 16), int(h[4:6], 16)
-        tc_is_light = (r + g + b) / 3 > 200
-    tokens["dark_mode"] = dark_signals or (dark_tw and not tc_is_light)
-
-    css_hrefs = re.findall(r'<link[^>]+href=["\']([^"\'>]+\.css[^"\'>]*)["\']', html, re.I)
-    if css_hrefs and base_url:
-        parsed = urlparse(base_url)
-        css_base = f"{parsed.scheme}://{parsed.netloc}"
-        total = len(css_hrefs)
-        seen: set[int] = set()
-        fetch_order: list[int] = []
-
-        # Front: first 4 (global setup: fonts, reset)
-        for i in range(min(4, total)):
-            seen.add(i)
-            fetch_order.append(i)
-        # Back: last 4 (late-loaded page-specific styles for bundled apps)
-        for i in range(total - 1, max(-1, total - 5), -1):
-            if i not in seen:
-                seen.add(i)
-                fetch_order.append(i)
-        # Middle: evenly-spaced samples across the full list
-        remaining_slots = 24 - len(fetch_order)
-        span = max(1, total - 8)
-        step = max(1, span // remaining_slots)
-        for i in range(4, total - 4, step):
-            if i not in seen and len(fetch_order) < 24:
-                seen.add(i)
-                fetch_order.append(i)
-
-        css_fetched = 0
-        max_fetch = 20
-        for idx in sorted(fetch_order):
-            if css_fetched >= max_fetch:
-                break
-            href = css_hrefs[idx]
-            css_url = href if href.startswith("http") else urljoin(css_base + "/", href.lstrip("/"))
-            try:
-                import urllib.request as _ureq
-                css_req = _ureq.Request(
-                    css_url,
-                    headers={"User-Agent": "ClutchDesign/1.0"},
-                )
-                with _ureq.urlopen(css_req, timeout=8.0) as cr:
-                    css_text = cr.read(300_000).decode("utf-8", errors="replace")
-                if re.search(r"@media\s*\(\s*prefers-color-scheme\s*:\s*dark\s*\)", css_text):
-                    tokens["dark_mode"] = True
-                for hc in re.findall(r"#([0-9a-fA-F]{6}|[0-9a-fA-F]{3})\b", css_text):
-                    norm = hc.upper()
-                    if len(norm) == 3:
-                        norm = norm[0]*2 + norm[1]*2 + norm[2]*2
-                    hex_counts["#" + norm] += 1
-                # Allow later definitions to override earlier ones (CSS cascade order)
-                for ck, cv in re.findall(r"--([\.\w-]+)\s*:\s*([^;}\n]{1,80})", css_text):
-                    css_vars[ck.strip()] = cv.strip().rstrip(",")
-                for rec in re.findall(r"var\(--([a-zA-Z][\w-]*)\)", css_text):
-                    css_var_refs.add(rec)
-                for ff in re.findall(r"font-family\s*:\s*([^;}\n]+)", css_text):
-                    cleaned = re.sub(r"['\"/]", "", ff.split(",")[0]).strip()
-                    if cleaned and cleaned not in font_families and len(cleaned) < 60:
-                        font_families.append(cleaned)
-                css_fetched += 1
-            except Exception:
-                continue
-        if hex_counts:
-            tokens["hex_colors"] = [h for h, _ in hex_counts.most_common(16)]
-            tokens["hex_colors_with_count"] = [[h, cnt] for h, cnt in hex_counts.most_common(16)]
-        if css_vars:
-            tokens["css_vars"] = css_vars
-        if css_var_refs:
-            tokens["css_var_references"] = sorted(css_var_refs)
-        if font_families:
-            tokens["font_families"] = font_families[:6]
-
-    return tokens
-
-
-def _format_css_tokens_for_prompt(tokens: dict[str, Any], host: str = "") -> str:
-    if not tokens:
-        return ""
-
-    lines: list[str] = ["[Extracted Design Tokens from Website]"]
-
-    theme_color = tokens.get("theme_color")
-    if theme_color:
-        # Skip near-black / near-white defaults — these are rarely the actual brand color
-        tc_is_default = False
-        if re.match(r'^#([0-9a-fA-F]{6}|[0-9a-fA-F]{3})$', theme_color):
-            h = theme_color.lstrip("#")
-            if len(h) == 3:
-                h = h[0]*2 + h[1]*2 + h[2]*2
-            r, g, b = int(h[0:2], 16), int(h[2:4], 16), int(h[4:6], 16)
-            brightness = (r + g + b) / 3
-            tc_is_default = brightness < 20 or brightness > 245
-        if not tc_is_default:
-            lines.append(f"Brand/theme color: {theme_color}")
-
-    hex_colors = tokens.get("hex_colors_with_count") or []
-    if hex_colors:
-        color_strs = [f"{h} (x{c})" for h, c in hex_colors[:10]]
-        lines.append(f"Hex colors found (by frequency): {', '.join(color_strs)}")
-
-    css_vars = tokens.get("css_vars") or {}
-    if css_vars:
-        var_strs = [f"--{k}: {v}" for k, v in list(css_vars.items())[:12]]
-        lines.append(f"CSS custom properties: {'; '.join(var_strs)}")
-
-    css_var_refs = tokens.get("css_var_references") or []
-    if css_var_refs:
-        ref_strs = [f"--{r}" for r in css_var_refs[:20]]
-        lines.append(f"CSS design token names referenced (hint at color palette): {', '.join(ref_strs)}")
-        # Extract color family hints from variable names
-        _skip_hints = {
-            "color", "primary", "secondary", "accent", "text", "bg", "background",
-            "muted", "dark", "alt", "neutral", "white", "black", "bold", "light",
-        }
-        color_hints = set()
-        for r in css_var_refs:
-            parts = r.lower().split("-")
-            if "color" in parts:
-                for p in parts[parts.index("color") + 1:]:
-                    if p not in _skip_hints:
-                        color_hints.add(p)
-        if color_hints:
-            lines.append(f"Color palette hints from variable names: {', '.join(sorted(color_hints))}")
-
-    tw_classes = tokens.get("tailwind_color_classes") or []
-    if tw_classes:
-        lines.append(f"Tailwind color classes used: {', '.join(tw_classes[:10])}")
-        tw_theme = tokens.get("tailwind_theme_color")
-        if tw_theme:
-            lines.append(f"Primary Tailwind color family: {tw_theme}-*")
-
-    font_families = tokens.get("font_families") or []
-    if font_families:
-        lines.append(f"Font families: {', '.join(font_families[:4])}")
-
-    dark_mode = tokens.get("dark_mode", False)
-    if dark_mode:
-        lines.append(
-            "Color mode: DARK (use dark background tones like #1a1a2e, #16213e, or soft dark grays — "
-            "avoid pure #000000 unless the site explicitly uses it)"
-        )
-        lines.append(
-            "CRITICAL: The reference website uses a DARK color scheme — the generated page MUST have a dark background."
-        )
-    else:
-        lines.append("Color mode: LIGHT (use white or very light backgrounds like #ffffff, #f8fafc, #fafafa)")
-        lines.append(
-            "CRITICAL: The reference website uses a LIGHT background — the generated page MUST have a white or very light background, with dark text. Do NOT generate a dark-themed page."
-        )
-
-    og_image = tokens.get("og_image")
-    if og_image:
-        lines.append(f"Reference screenshot (OG image): {og_image}")
-
-    if len(lines) > 1:
-        lines.append(
-            "\nCRITICAL: You MUST use these exact colors and fonts in your design spec. "
-            "Do NOT invent or substitute colors — the spec must faithfully match the source website's palette."
-        )
-
-    return "\n".join(lines) if len(lines) > 1 else ""
-
-
-def _fetch_url_snapshot(url: str) -> dict[str, Any]:
-    import urllib.error
-    import urllib.request
-    from html import unescape
-    from urllib.parse import urlparse
-
-    req = urllib.request.Request(
-        url,
-        headers={
-            "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 ClutchDesign/1.0",
-            "Accept": "text/html,application/xhtml+xml;q=0.9,*/*;q=0.8",
-            "Accept-Language": "en-US,en;q=0.9",
-        },
-        method="GET",
-    )
-    try:
-        with urllib.request.urlopen(req, timeout=12.0) as resp:
-            raw = resp.read(500_000)
-            charset = "utf-8"
-            ctype = resp.headers.get_content_charset()
-            if ctype:
-                charset = ctype
-            html = raw.decode(charset, errors="replace")
-            final_url = resp.geturl() or url
-    except urllib.error.HTTPError as exc:
-        raise DesignError(f"Could not fetch URL ({exc.code})") from exc
-    except Exception as exc:
-        raise DesignError(f"Could not fetch URL: {exc}") from exc
-
-    title_m = re.search(r"<title[^>]*>(.*?)</title>", html, re.I | re.S)
-    title = unescape(re.sub(r"\s+", " ", title_m.group(1))).strip() if title_m else ""
-    desc_m = re.search(
-        r'<meta[^>]+name=["\']description["\'][^>]+content=["\']([^"\'>]+)',
-        html,
-        re.I | re.S,
-    ) or re.search(
-        r'<meta[^>]+content=["\']([^"\'>]+)["\'][^>]+name=["\']description',
-        html,
-        re.I | re.S,
-    )
-    description = unescape(re.sub(r"\s+", " ", desc_m.group(1))).strip() if desc_m else ""
-    cleaned = re.sub(r"(?is)<(script|style|noscript)[^>]*>.*?</\1>", " ", html)
-    cleaned = re.sub(r"(?is)<[^>]+>", " ", cleaned)
-    cleaned = unescape(re.sub(r"\s+", " ", cleaned)).strip()
-    excerpt = cleaned[:4000]
-    host = urlparse(final_url).netloc or urlparse(url).netloc
-
-    try:
-        css_tokens = _extract_css_tokens(html, base_url=final_url)
-    except Exception as exc:
-        logger.debug("design url css_tokens extraction failed url=%s err=%s", url, exc)
-        css_tokens = {}
-
-    og_image_analysis: dict[str, Any] = {}
-    og_img_url = css_tokens.get("og_image")
-    if og_img_url:
-        try:
-            import base64 as _b64
-            import io as _io
-            og_req = urllib.request.Request(
-                og_img_url,
-                headers={"User-Agent": "ClutchDesign/1.0"},
-            )
-            with urllib.request.urlopen(og_req, timeout=10.0) as img_resp:
-                img_bytes = img_resp.read(2_000_000)
-            img_b64 = _b64.b64encode(img_bytes).decode("ascii")
-            content_type = img_resp.headers.get_content_type() or "image/png"
-            data_url = f"data:{content_type};base64,{img_b64}"
-            from src.design.image_analysis import analyze_image_for_spec
-            og_image_analysis = analyze_image_for_spec(data_url)
-            logger.info(
-                "design url OG image analyzed url=%s colors=%d",
-                og_img_url, len(og_image_analysis.get("colors", [])),
-            )
-        except Exception as exc:
-            logger.debug("design url OG image analysis failed url=%s err=%s", og_img_url, exc)
-
-    # Browser-based extraction: render the page, extract computed styles + screenshot
-    browser_tokens: dict[str, Any] = {}
-    try:
-        from src.design.browser_extract import extract_website_tokens
-        browser_tokens = extract_website_tokens(final_url, timeout_sec=40.0)
-        if browser_tokens.get("available"):
-            logger.info(
-                "design url browser extraction succeeded url=%s fonts=%d",
-                final_url, len((browser_tokens.get("computed_styles") or {}).get("fonts", [])),
-            )
-        else:
-            logger.debug(
-                "design url browser extraction unavailable url=%s err=%s",
-                final_url, browser_tokens.get("error", "unknown"),
-            )
-    except Exception as exc:
-        logger.debug("design url browser extraction failed url=%s err=%s", final_url, exc)
-
-    return {
-        "url": final_url,
-        "host": host,
-        "title": title[:200],
-        "description": description[:400],
-        "excerpt": excerpt,
-        "css_tokens": css_tokens,
-        "og_image_analysis": og_image_analysis,
-        "browser_computed_styles": browser_tokens.get("computed_styles") or {},
-        "browser_screenshot": browser_tokens.get("screenshot_data_url") or "",
-        "browser_prompt_fragment": browser_tokens.get("prompt_fragment") or "",
-        "fetched_at": now_iso(),
-    }
-
-
-def _build_ui_compact_prompt(
-    *,
-    user_prompt: str,
-    spec: dict[str, Any],
-    device: str,
-) -> str:
-    pattern = str(spec.get("layout_pattern") or detect_layout_pattern(user_prompt, device=device))
-    primary = _first_hex((spec.get("colors") or {}), "primary", "#171717")
-    return (
-        "Generate ONE complete HTML page with Tailwind CDN for this brief.\n"
-        f"Brief: {user_prompt}\n"
-        f"Device: {device}\n"
-        f"Layout: {pattern}\n"
-        f"Primary color: {primary}\n"
-        f"Brand: {spec.get('name') or 'Clutch'}\n"
-        "Rules: full <!DOCTYPE html>…</html>, visible UI in body, match the brief intent.\n"
-        "Prefer semantic <button> / <a role=\"button\"> with data-clutch-id for IUE; "
-        "do NOT add real navigation (href=\"#\" or javascript:) or <form> submit / onclick handlers.\n"
-        "Return ONLY ```html ... ```."
-    )
-
-
-def _build_shared_layout_context(screens_to_gen: list[dict[str, Any]]) -> str:
-    """Build a shared layout context string for multi-page generation.
-    
-    Tells the LLM that all non-login pages must share a consistent left sidebar
-    with identical navigation items, colors, and ordering.
-    """
-    if len(screens_to_gen) <= 1:
-        return ""
-    login_keywords = ("login", "signin", "sign-in", "auth", "登", "注册")
-    login_ids = set()
-    for s in screens_to_gen:
-        pid = (s.get("prompt") or "").lower()
-        nid = (s.get("name") or "").lower()
-        if any(kw in pid or kw in nid for kw in login_keywords):
-            login_ids.add(s["id"])
-    non_login = [s for s in screens_to_gen if s["id"] not in login_ids]
-    if len(non_login) <= 1:
-        return ""
-    nav_names = [s["name"] for s in non_login]
-    lines = [
-        f"This is a multi-page prototype with {len(non_login)} non-login screens plus {len(login_ids)} login/auth screens.",
-        "",
-        "LAYOUT RULES:",
-        f"- ALL {len(non_login)} non-login pages ({', '.join(nav_names)}) MUST share an identical left sidebar navigation.",
-        "- Sidebar bg color: use the design system's primary or secondary color (consistent across all pages).",
-        "- Sidebar text color: white or on-primary (consistent across all pages).",
-        f"- Navigation items in exact order: {' > '.join(nav_names)}.",
-        "- Every page must highlight its own nav item as active/selected.",
-        "- Sidebar width, padding, font sizes, and spacing must be pixel-identical across all pages.",
-        "- Login/auth pages: centered card layout WITHOUT sidebar (standalone authentication screen).",
-        "",
-        "These rules ensure visual consistency when users navigate between screens.",
-    ]
-    return "\n".join(lines)
-
-
-def _build_ui_generation_prompt(
-    *,
-    user_prompt: str,
-    spec: dict[str, Any],
-    device: str,
-    pattern: str,
-    design_md: str = "",
-    md_text: str | None = None,
-    url_snapshot: dict[str, Any] | None = None,
-    has_image: bool = False,
-    image_attached: bool = False,
-    current_html: str = "",
-    instruction: str = "",
-    shared_layout_context: str = "",
-) -> str:
-    fewshot = fewshot_for_pattern(pattern)
-    layout_hint = layout_wrapper_hint(pattern)
-    ui_parts = [
-        "You are an expert web designer. Generate a single, fully-formed, beautiful HTML document "
-        "using Tailwind CSS.\n",
-        f"Brief: {user_prompt}\n",
-    ]
-    if instruction:
-        ui_parts.append(f"Revision instruction: {instruction}\n")
-    ui_parts.append(
-        (
-            "Device target: mobile app viewport 390×844 CSS pixels (iPhone-class). "
-            "Use a single-column stacked layout, large touch targets, no desktop sidebar.\n"
-            if device == "app"
-            else
-            "Device target: desktop web viewport 1920×1080 CSS pixels (16:9). "
-            "Use a full-width desktop layout (top nav, multi-column grids, wide hero). "
-            "Do NOT produce a narrow phone-only page.\n"
-        )
-    )
-    intent = _prompt_intent(user_prompt)
-    if intent == "login":
-        intent_rule = (
-            "5. Brief intent: login — build a polished sign-in / registration screen.\n"
-        )
-    else:
-        intent_rule = (
-            f"5. Brief intent: {intent} — the page MUST match this intent; "
-            "do NOT default to a generic login/auth screen unless the brief asks for it.\n"
-        )
-    ui_parts += [
-        f"Layout pattern: {pattern}\n",
-        f"Layout constraints: {layout_hint}\n",
-    ]
-    if shared_layout_context:
-        ui_parts.append(
-            "\n=== SHARED PAGE LAYOUT (applies to ALL screens in this multi-page session) ===\n"
-            + shared_layout_context + "\n"
-            "=== END SHARED LAYOUT ===\n\n"
-        )
-    ui_parts += [
-        (
-            "CRITICAL Rules:\n"
-            "1. Use Tailwind CDN: <script src=\"https://cdn.tailwindcss.com\"></script>\n"
-            "2. Define design system colors via tailwind.config script — no custom <style> blocks.\n"
-            "3. Complete HTML with closed </html>; keep body under 80 lines to avoid truncation.\n"
-            "4. Implement the screen in the Brief — match the requested screen type.\n"
-        )
-        + intent_rule
-        + (
-            "6. COLOR RULES — VIOLATION CAUSES REJECTION:\n"
-            "   a. Use ONLY color hex codes from the Design system JSON. NEVER invent arbitrary hex values like #1A73E8 or #3B82F6.\n"
-            "   b. Do NOT use Tailwind default color names: NO bg-blue-*, bg-red-*, bg-green-*, bg-yellow-*, text-blue-*, etc.\n"
-            "   c. Only use the custom color palette defined in the tailwind.config (e.g. bg-primary, bg-accent, bg-secondary, text-primary, bg-neutral). Reference the Design system JSON colors.\n"
-            "   d. Page background MUST be solid white or the neutral/surface color from the design system. NO colorful gradient backgrounds like bg-gradient-to-br from-[...].\n"
-            "   e. Primary buttons: bg-primary text-white rounded-xl. Secondary buttons: border-gray-300 text-gray-700 bg-white. Accent actions: bg-accent text-white.\n"
-            "7. Select 3-5 core components; use rounded-2xl cards, professional horizontal padding px-6–px-10, hover opacity transitions only.\n"
-            "8. Prototype interactivity: use semantic <button> / <a role=\"button\"> with data-clutch-id "
-            "so IUE can detect hotspots; no real navigation hrefs, no <form> submit, no onclick/javascript.\n"
-            f"9. {_TASTE_ANTI_PATTERNS}"
-        ),
-        f"Style reference (match quality & polish only — do NOT copy this layout verbatim):\n{fewshot}\n",
-        f"Design system JSON:\n{json.dumps(spec, ensure_ascii=False)}\n",
-    ]
-    if design_md:
-        cap = 2000 if str(spec.get("name") or "").strip().lower() == "clutch" else 8000
-        ui_parts.append(f"DESIGN.md rules:\n{design_md[:cap]}\n")
-    if md_text:
-        ui_parts.append(
-            f"\n=== MANDATORY DESIGN SPECIFICATION ({md_text[:60]!r}...) ===\n"
-            f"{md_text[:8000]}\n"
-            "=== END MANDATORY SPECIFICATION ===\n"
-            "CRITICAL COMPLIANCE RULES for the above specification:\n"
-            "- Use ONLY the exact color hex codes specified — NEVER substitute or invent colors.\n"
-            "- Use ONLY the font families named in the specification — load via Google Fonts @import if needed.\n"
-            "- Apply all spacing, border-radius, and shadow values exactly as specified.\n"
-            "- Implement every component mentioned in the specification.\n"
-            "- The generated HTML must be a faithful implementation of this spec. No deviations.\n"
-        )
-    if url_snapshot:
-        browser_frag = url_snapshot.get("browser_prompt_fragment") or ""
-        if browser_frag:
-            ui_parts.append(
-                f"Reference website: {url_snapshot.get('url')} "
-                f"({url_snapshot.get('title') or url_snapshot.get('host')}).\n"
-                + browser_frag + "\n"
-                "Apply the above design tokens faithfully in your HTML output.\n"
-                "Determine the background color from the Body computed styles or CSS custom properties. "
-                "If the reference body background is transparent, white, or light (<#ccc), you MUST use a light background (#ffffff or similar). "
-                "If it's dark (>#333), use a dark background. Match what the reference actually renders.\n"
-            )
-        else:
-            css_tokens = url_snapshot.get("css_tokens") or {}
-            token_desc = _format_css_tokens_for_prompt(css_tokens, host=str(url_snapshot.get("host") or ""))
-            if token_desc:
-                ui_parts.append(
-                    f"Reference website: {url_snapshot.get('url')} "
-                    f"({url_snapshot.get('title') or url_snapshot.get('host')}).\n"
-                    + token_desc + "\n"
-                    "Apply the above design tokens faithfully in your HTML output.\n"
-                )
-            else:
-                ui_parts.append(
-                    f"Visual inspiration from {url_snapshot.get('url')} "
-                    f"({url_snapshot.get('title') or url_snapshot.get('host')}).\n"
-                )
-    if image_attached:
-        ui_parts.append(
-            "Match the attached reference screenshot (structure, hierarchy, spacing).\n"
-        )
-    if current_html:
-        ui_parts.append(
-            "=== REVISION MODE — CRITICAL ===\n"
-            "You are REVISING an existing page, NOT creating a new one from scratch.\n"
-            "1. KEEP the existing page layout, structure, and content type unchanged.\n"
-            "2. Only apply the specific changes requested in the Revision instruction above.\n"
-            "3. Do NOT replace the page with a different screen type (e.g., don't turn a dashboard into a login form).\n"
-            "4. The page type and purpose must remain the same as the current HTML below.\n"
-            f"Current HTML to revise:\n{current_html[:14000]}\n"
-        )
-    ui_parts.append("Return ONLY the HTML document inside ```html ... ```.")
-    return "".join(ui_parts)
-
-
-def _design_review_and_improve(
-    router: Any,
-    *,
-    html: str,
-    spec: dict[str, Any],
-    user_prompt: str,
-    model_id: str,
-    device: str,
-) -> tuple[str, str | None, int, dict[str, int], bool]:
-    """Design Review Pass: score HTML; refine once if below threshold."""
-    review_prompt = (
-        "You are a senior UI design reviewer. Score this HTML mockup 1-10 on:\n"
-        "- Visual hierarchy & modern aesthetics\n"
-        "- Spacing & typography consistency\n"
-        "- Color consistency & contrast\n"
-        "- CTA clarity & accessibility\n"
-        "- Responsive layout\n\n"
-        f"Brief: {user_prompt}\nDevice: {device}\n"
-        f"Design system: {json.dumps(spec, ensure_ascii=False)[:4000]}\n\n"
-        f"HTML:\n{html[:12000]}\n\n"
-        'Return ONLY JSON: {"score": N, "feedback": "..."}'
-    )
-    review_text, review_reasoning, review_usage, review_estimated = _llm_complete(
-        router, review_prompt, model_id=model_id
-    )
-    score, feedback = parse_review_score(review_text)
-    combined_reasoning = review_reasoning
-    usage = review_usage
-    estimated = review_estimated
-    if score >= review_threshold():
-        return html, combined_reasoning, score, usage, estimated
-    improve_prompt = (
-        "Improve this HTML UI based on the design review feedback. "
-        "Apply concrete fixes to spacing, hierarchy, CTAs, and contrast.\n"
-        "Keep semantic buttons with data-clutch-id; no real navigation / form submit / onclick handlers.\n"
-        f"Feedback (score {score}/10): {feedback}\n"
-        f"Brief: {user_prompt}\n"
-        f"HTML:\n{html[:14000]}\n"
-        "Use Tailwind CDN + tailwind.config for colors. Return ONLY ```html ... ```."
-    )
-    improved_text, improve_reasoning, improve_usage, improve_estimated = _llm_complete(
-        router, improve_prompt, model_id=model_id
-    )
-    improved = _extract_html_from_llm(improved_text)
-    usage = merge_token_usage(usage, improve_usage)
-    estimated = estimated or improve_estimated
-    if improve_reasoning:
-        combined_reasoning = "\n---\n".join(filter(None, [combined_reasoning, improve_reasoning]))
-    if _html_has_visible_content(improved):
-        return improved, combined_reasoning, score, usage, estimated
-    return html, combined_reasoning, score, usage, estimated
-
-
-def _build_ui_correction_prompt(
-    *,
-    user_prompt: str,
-    spec: dict[str, Any],
-    device: str,
-    reason: str,
-) -> str:
-    layout_hint = layout_wrapper_hint(
-        str(spec.get("layout_pattern") or detect_layout_pattern(user_prompt, device=device))
-    )
-    return (
-        "You are an expert web designer. Your previous HTML was rejected.\n"
-        f"Rejection reason: {reason}\n"
-        f"Brief (follow exactly): {user_prompt}\n"
-        f"Layout constraints: {layout_hint}\n"
-        f"Design system JSON:\n{json.dumps(spec, ensure_ascii=False)[:6000]}\n"
-        "Rules:\n"
-        "- Generate a NEW complete HTML document using Tailwind CDN.\n"
-        "- Fulfill the brief — do NOT default to login/auth unless explicitly requested.\n"
-        "- Use colors and typography from the design system; make it visually distinct.\n"
-        "- Prefer semantic <button> with data-clutch-id for IUE; no real navigation / form submit / onclick.\n"
-        f"- {_TASTE_ANTI_PATTERNS}"
-        "Return ONLY ```html ... ```."
-    )
-
-
-def _generate_ui_html(
-    router: Any,
-    *,
-    user_prompt: str,
-    spec: dict[str, Any],
-    device: str,
-    model_id: str,
-    design_md: str = "",
-    md_text: str | None = None,
-    url_snapshot: dict[str, Any] | None = None,
-    has_image: bool = False,
-    image_data_url: str | None = None,
-    current_html: str = "",
-    instruction: str = "",
-    fallback_html: str | None = None,
-    shared_layout_context: str = "",
-) -> tuple[str, str | None, dict[str, int], bool, str | None]:
-    last_fail: str | None = None
-    pattern = str(spec.get("layout_pattern") or detect_layout_pattern(user_prompt, device=device))
-    meta = _build_ui_generation_prompt(
-        user_prompt=user_prompt,
-        spec=spec,
-        device=device,
-        pattern=pattern,
-        design_md=design_md,
-        md_text=md_text,
-        url_snapshot=url_snapshot,
-        has_image=has_image,
-        image_attached=bool(image_data_url),
-        current_html=current_html,
-        instruction=instruction,
-        shared_layout_context=shared_layout_context,
-    )
-    text, reasoning, usage, estimated, fail = _try_llm_complete_vision(
-        router, meta, model_id=model_id, image_data_url=image_data_url
-    )
-    if fail:
-        last_fail = fail
-    usage_acc = usage
-    estimated_acc = estimated
-    raw = _html_from_llm_response(text, reasoning)
-    html = _coerce_ui_html(
-        raw,
-        title=str(spec.get("name") or "UI"),
-        prompt=user_prompt,
-        spec=spec,
-        device=device,
-        fallback_html=fallback_html,
-        allow_template_fallback=False,
-    )
-    # If the HTML itself contains a vision error message, drop image and retry immediately
-    if _html_has_visible_content(html) and image_data_url and _looks_like_vision_error(html):
-        logger.warning("design ui HTML contains vision error — dropping image, retrying text-only")
-        no_img_meta = _build_ui_generation_prompt(
-            user_prompt=user_prompt,
-            spec=spec,
-            device=device,
-            pattern=pattern,
-            design_md=design_md,
-            md_text=md_text,
-            url_snapshot=url_snapshot,
-            has_image=False,
-            image_attached=False,
-            current_html=current_html,
-            instruction=instruction,
-            shared_layout_context=shared_layout_context,
-        )
-        ni_text, ni_reasoning, ni_usage, ni_estimated, ni_fail = _try_llm_complete_vision(
-            router, no_img_meta, model_id=model_id, image_data_url=None,
-        )
-        usage_acc = merge_token_usage(usage_acc, ni_usage)
-        estimated_acc = estimated_acc or ni_estimated
-        ni_raw = _html_from_llm_response(ni_text, ni_reasoning)
-        ni_html = _coerce_ui_html(
-            ni_raw,
-            title=str(spec.get("name") or "UI"),
-            prompt=user_prompt,
-            spec=spec,
-            device=device,
-            fallback_html=fallback_html,
-            allow_template_fallback=True,
-        )
-        if _html_has_visible_content(ni_html):
-            html = ni_html
-            raw = ni_raw
-            text = ni_text
-            reasoning = ni_reasoning
-            last_fail = None
-    if _html_has_visible_content(html) and not _html_matches_brief_intent(user_prompt, html):
-        detected = _detect_html_intent(html)
-        expected = _prompt_intent(user_prompt)
-        logger.warning(
-            "design ui intent mismatch prompt=%r detected=%s expected=%s — LLM retry",
-            user_prompt[:80],
-            detected,
-            expected,
-        )
-        correction = _build_ui_correction_prompt(
-            user_prompt=user_prompt,
-            spec=spec,
-            device=device,
-            reason=(
-                f"Page looks like '{detected}' but brief requires '{expected}'. "
-                "Do not output login/auth unless the brief asks for it."
-            ),
-        )
-        retry_text, retry_reasoning, retry_usage, retry_estimated, fail = _try_llm_complete_vision(
-            router, correction, model_id=model_id, image_data_url=image_data_url
-        )
-        if fail:
-            last_fail = fail
-        usage_acc = merge_token_usage(usage_acc, retry_usage)
-        estimated_acc = estimated_acc or retry_estimated
-        retry_raw = _html_from_llm_response(retry_text, retry_reasoning)
-        retry_html = _coerce_ui_html(
-            retry_raw,
-            title=str(spec.get("name") or "UI"),
-            prompt=user_prompt,
-            spec=spec,
-            device=device,
-            fallback_html=html,
-            allow_template_fallback=False,
-        )
-        if _html_has_visible_content(retry_html):
-            html = retry_html
-            reasoning = "\n---\n".join(filter(None, [reasoning, retry_reasoning]))
-    elif not _html_has_visible_content(html):
-        logger.warning("design ui blank output prompt=%r — compact LLM retry", user_prompt[:80])
-        compact = _build_ui_compact_prompt(user_prompt=user_prompt, spec=spec, device=device)
-        retry_text, retry_reasoning, retry_usage, retry_estimated, fail = _try_llm_complete_vision(
-            router, compact, model_id=model_id, image_data_url=image_data_url
-        )
-        if fail:
-            last_fail = fail
-        usage_acc = merge_token_usage(usage_acc, retry_usage)
-        estimated_acc = estimated_acc or retry_estimated
-        retry_raw = _html_from_llm_response(retry_text, retry_reasoning)
-        retry_html = _coerce_ui_html(
-            retry_raw,
-            title=str(spec.get("name") or "UI"),
-            prompt=user_prompt,
-            spec=spec,
-            device=device,
-            fallback_html=fallback_html,
-            allow_template_fallback=False,
-        )
-        if _html_has_visible_content(retry_html):
-            html = retry_html
-            reasoning = "\n---\n".join(filter(None, [reasoning, retry_reasoning]))
-        else:
-            retry_meta = (
-                meta
-                + "\n\nIMPORTANT: Your previous response was empty or invalid. "
-                "Output a complete, visible HTML document that fulfills the brief."
-            )
-            retry2_text, retry2_reasoning, retry2_usage, retry2_estimated, fail = _try_llm_complete_vision(
-                router, retry_meta, model_id=model_id, image_data_url=image_data_url
-            )
-            if fail:
-                last_fail = fail
-            usage_acc = merge_token_usage(usage_acc, retry2_usage)
-            estimated_acc = estimated_acc or retry2_estimated
-            retry2_raw = _html_from_llm_response(retry2_text, retry2_reasoning)
-            retry2_html = _coerce_ui_html(
-                retry2_raw,
-                title=str(spec.get("name") or "UI"),
-                prompt=user_prompt,
-                spec=spec,
-                device=device,
-                fallback_html=fallback_html,
-                allow_template_fallback=False,
-            )
-            if _html_has_visible_content(retry2_html):
-                html = retry2_html
-                reasoning = "\n---\n".join(
-                    filter(None, [reasoning, retry_reasoning, retry2_reasoning])
-                )
-    # --- Final degradation: if image was attached but all retries failed, drop image and retry ---
-    if not _html_has_visible_content(html) and image_data_url:
-        logger.warning("design ui all retries failed with image — dropping image, retrying text-only")
-        fallback_meta = _build_ui_generation_prompt(
-            user_prompt=user_prompt,
-            spec=spec,
-            device=device,
-            pattern=pattern,
-            design_md=design_md,
-            md_text=md_text,
-            url_snapshot=url_snapshot,
-            has_image=False,
-            image_attached=False,
-            current_html=current_html,
-            instruction=instruction,
-            shared_layout_context=shared_layout_context,
-        )
-        fb_text, fb_reasoning, fb_usage, fb_estimated, fb_fail = _try_llm_complete_vision(
-            router, fallback_meta, model_id=model_id, image_data_url=None,
-        )
-        usage_acc = merge_token_usage(usage_acc, fb_usage)
-        estimated_acc = estimated_acc or fb_estimated
-        fb_raw = _html_from_llm_response(fb_text, fb_reasoning)
-        fb_html = _coerce_ui_html(
-            fb_raw,
-            title=str(spec.get("name") or "UI"),
-            prompt=user_prompt,
-            spec=spec,
-            device=device,
-            fallback_html=fallback_html,
-            allow_template_fallback=True,
-        )
-        if _html_has_visible_content(fb_html):
-            html = fb_html
-            reasoning = "\n---\n".join(filter(None, [reasoning, fb_reasoning]))
-            last_fail = None
-            logger.info("design ui text-only fallback succeeded after image failure")
-    if _html_has_visible_content(html) and _DESIGN_REVIEW_ENABLED:
-        reviewed, review_reasoning, _score, review_usage, review_estimated = _design_review_and_improve(
-            router,
-            html=html,
-            spec=spec,
-            user_prompt=user_prompt,
-            model_id=model_id,
-            device=device,
-        )
-        usage_acc = merge_token_usage(usage_acc, review_usage)
-        estimated_acc = estimated_acc or review_estimated
-        html = _coerce_ui_html(
-            reviewed,
-            title=str(spec.get("name") or "UI"),
-            prompt=user_prompt,
-            spec=spec,
-            device=device,
-            fallback_html=html,
-            allow_template_fallback=False,
-        )
-        if review_reasoning:
-            reasoning = "\n---\n".join(filter(None, [reasoning, review_reasoning]))
-    if not _html_has_visible_content(html) and not last_fail:
-        last_fail = "the model returned no valid HTML"
-    return html, reasoning, usage_acc, estimated_acc, last_fail
 
 
 def generate_session(
@@ -3023,657 +1257,3 @@ def start_iterate_session(
     return public_session_payload(manifest, sdir)
 
 
-def _html_essentially_same(a: str, b: str) -> bool:
-    def norm(s: str) -> str:
-        out = re.sub(r"\s+", "", s or "")
-        out = re.sub(r'data-note="[^"]*"', "", out, flags=re.I)
-        return out
-    return bool(a) and bool(b) and norm(a) == norm(b)
-
-
-def _merged_design_prompt(manifest: dict[str, Any], instruction: str) -> str:
-    base = str(manifest.get("prompt") or manifest.get("name") or "").strip()
-    note = (instruction or "").strip()
-    if base and note:
-        return f"{base}\n{note}"
-    return note or base or "Interface"
-
-
-def _infer_iterate_mode(instruction: str, *, mode: str | None, target_kind: str | None) -> str:
-    raw = (mode or "auto").strip().lower()
-    if raw in {"modify", "add", "duplicate", "variant", "revise_spec"}:
-        return raw
-    text = instruction.lower()
-    if any(k in text for k in ("variant", "变体", "换个布局", "another version", "explore layout")):
-        return "variant"
-    if any(k in text for k in ("revise spec", "改规范", "更新设计系统", "update design system")):
-        return "revise_spec"
-    add_keys = (
-        "新增",
-        "添加一",
-        "再做",
-        "另一个",
-        "新页面",
-        "新画板",
-        "再来一",
-        "add ",
-        "new page",
-        "another ",
-        "create a new",
-        "also create",
-        "new screen",
-        "new artboard",
-    )
-    mod_keys = (
-        "改成",
-        "修改",
-        "优化",
-        "调整",
-        "换成",
-        "改一下",
-        "要体现",
-        "需要",
-        "显示",
-        "加上",
-        "增加",
-        "完善",
-        "深色",
-        "fix",
-        "change ",
-        "update ",
-        "make it",
-        "dark mode",
-        "improve",
-        "tweak",
-        "refine",
-        "add lyrics",
-        "show ",
-    )
-    has_add = any(k in text for k in add_keys)
-    has_mod = any(k in text for k in mod_keys)
-    if has_add and not has_mod:
-        return "add"
-    if has_mod and not has_add:
-        return "modify"
-    if has_mod and has_add and target_kind == "ui":
-        return "modify"
-    if target_kind == "ui":
-        return "modify"
-    return "add"
-
-
-def _next_screen_id(screens: list[dict[str, Any]]) -> str:
-    used = {str(s.get("id") or "") for s in screens}
-    if "main" not in used:
-        return "main"
-    i = 2
-    while f"screen-{i}" in used:
-        i += 1
-    return f"screen-{i}"
-
-
-def _screen_layout_x(
-    screens: list[dict[str, Any]], *, device: str = "web", has_source: bool = False
-) -> int:
-    xs = []
-    for s in screens:
-        pos = s.get("position") or {}
-        if isinstance(pos, dict) and isinstance(pos.get("x"), (int, float)):
-            xs.append(int(pos["x"]))
-    step = ui_layout_step(device)
-    return (max(xs) + step) if xs else default_ui_origin_x(has_source=has_source)
-
-
-def iterate_session(
-    run_id: str,
-    instruction: str,
-    *,
-    target_kind: str | None = None,
-    target_id: str | None = None,
-    element_path: str | None = None,
-    element_label: str | None = None,
-    mode: str | None = None,
-) -> dict[str, Any]:
-    from src.models_config import get_router, is_model_available
-
-    sdir = session_dir(run_id)
-    manifest = read_manifest(sdir)
-    instruction = instruction.strip()
-    if not instruction:
-        raise DesignError("Instruction is required")
-    screens = [s for s in (manifest.get("screens") or []) if not s.get("deleted")]
-    if not screens and (target_kind or "ui") == "ui":
-        raise DesignError("Generate a design before iterating")
-
-    kind = (target_kind or "ui").strip().lower()
-    if kind not in {"ui", "spec", "md", "image", "url", "process"}:
-        kind = "ui"
-    action = _infer_iterate_mode(instruction, mode=mode, target_kind=kind)
-    if action == "revise_spec":
-        kind = "spec"
-        action = "modify"
-    elif action == "variant":
-        kind = "ui"
-        action = "add"
-        instruction = (
-            "Generate a design variant exploring layout / color / content aspects. "
-            f"Direction: {instruction}"
-        )
-    design_md = (sdir / DESIGN_MD).read_text(encoding="utf-8") if (sdir / DESIGN_MD).is_file() else ""
-    spec = manifest.get("spec")
-    if not isinstance(spec, dict) and (sdir / SPEC_JSON).is_file():
-        try:
-            spec = json.loads((sdir / SPEC_JSON).read_text(encoding="utf-8"))
-        except json.JSONDecodeError:
-            spec = None
-
-    selection_note = f"Selected: {kind}"
-    if target_id:
-        selection_note += f"/{target_id}"
-    if element_label or element_path:
-        selection_note += f"; element={element_label or element_path}"
-
-    log = list(manifest.get("process_log") or [])
-    log_start = len(log)
-    log.append(
-        {
-            "role": "user",
-            "text": f"{instruction} [{selection_note}; mode={action}]",
-            "at": now_iso(),
-        }
-    )
-    log.append(
-        {
-            "role": "assistant",
-            "text": (
-                "Creating a new version…"
-                if action == "add" and kind == "ui"
-                else "Thinking… applying your changes to the selected design."
-            ),
-            "status": "iterating",
-            "at": now_iso(),
-        }
-    )
-    manifest["status"] = "iterating"
-    manifest["process_log"] = log
-    write_manifest(sdir, manifest)
-
-    router = get_router()
-    model_id = router.active_model_id
-    model_id, model_name = _stamp_session_model(
-        manifest,
-        router,
-        model_id=model_id,
-        process_log=log,
-        record_in_log=False,
-    )
-    for i in range(len(log) - 1, -1, -1):
-        if log[i].get("role") == "assistant" and log[i].get("status") == "iterating":
-            log[i] = {
-                **log[i],
-                "model_id": model_id,
-                "model_name": model_name,
-            }
-            break
-    manifest["process_log"] = log
-    write_manifest(sdir, manifest)
-
-    if kind in {"spec", "md"}:
-        updated_spec = spec if isinstance(spec, dict) else _fallback_spec(instruction)
-        spec_usage = empty_token_usage()
-        spec_usage_estimated = False
-        if is_model_available(router, model_id):
-            try:
-                meta = (
-                    "You revise a product design system JSON.\n"
-                    f"Instruction: {instruction}\n"
-                    f"Current design system JSON:\n{json.dumps(updated_spec, ensure_ascii=False)}\n"
-                    f"Source DESIGN.md (excerpt):\n{design_md[:8000]}\n"
-                    "Return ONLY updated JSON with keys: name, rationale, colors, typography, components."
-                )
-                spec_raw, _, spec_usage, spec_usage_estimated = _llm_complete(
-                    router, meta, model_id=model_id
-                )
-                parsed = _extract_json_block(spec_raw)
-                if isinstance(parsed, dict):
-                    updated_spec = enrich_fallback_spec(
-                        parsed,
-                        instruction,
-                        str(parsed.get("layout_pattern") or detect_layout_pattern(instruction)),
-                    )
-            except Exception as exc:
-                logger.warning("design iterate spec failed: %s", exc)
-        (sdir / SPEC_JSON).write_text(
-            json.dumps(updated_spec, indent=2, ensure_ascii=False) + "\n", encoding="utf-8"
-        )
-        if kind == "md" or not design_md:
-            (sdir / DESIGN_MD).write_text(_spec_to_design_md(updated_spec), encoding="utf-8")
-        manifest["spec"] = updated_spec
-        clear_fake_thumbnail(sdir)
-        manifest.pop("thumbnail", None)
-        spec_updated_text = "Design system updated."
-        if not is_model_available(router, model_id):
-            spec_updated_text += f"\n\n⚠️ Warning: Model '{model_id}' is not available (API key missing in Settings -> Models). Using offline fallback templates."
-        _finalize_assistant_step(
-            log,
-            text=spec_updated_text,
-            status="ready",
-            model_id=model_id,
-            model_name=model_name,
-            usage=spec_usage,
-            usage_estimated=spec_usage_estimated,
-            replace_statuses={"iterating"},
-        )
-        manifest["process_log"] = log
-        manifest["status"] = "ready"
-        write_manifest(sdir, manifest)
-        return public_session_payload(manifest, sdir)
-
-    if not screens:
-        raise DesignError("Generate a design before iterating")
-
-    ref_rel = manifest.get("reference_image")
-    image_data_url = load_reference_data_url(sdir, str(ref_rel) if ref_rel else None)
-    has_image = bool(image_data_url)
-
-    md_rel = manifest.get("reference_md")
-    md_text, md_name = load_reference_md(sdir, str(md_rel) if md_rel else None)
-    has_md = bool(md_text)
-
-    url = manifest.get("reference_url")
-    url_snapshot = load_url_snapshot(sdir) if url else None
-    has_url = bool(url_snapshot)
-
-    ui_vision_ok = _check_vision_ok(router, model_id, image_data_url)
-
-    device = str(manifest.get("device") or "web")
-
-    if action == "add":
-        spec_dict = spec if isinstance(spec, dict) else _fallback_spec(instruction)
-        screens_to_gen = _parse_multi_screens(instruction, model_id, router)
-        
-        new_screens = []
-        ui_origin_x = _screen_layout_x(screens, device=device, has_source=has_md or has_image or has_url)
-        for idx, screen_info in enumerate(screens_to_gen):
-            sid = _next_screen_id(screens + new_screens)
-            sname = screen_info["name"]
-            step = ui_layout_step(device)
-            x_pos = ui_origin_x + idx * step
-            new_screens.append({
-                "id": sid,
-                "name": sname,
-                "position": {"x": x_pos, "y": _DESIGN_ROW_Y},
-                "html_path": None,
-                "active_round_index": 0,
-                "original_info": screen_info,
-            })
-            
-        # Append new screens to manifest immediately as placeholders
-        manifest["screens"] = screens + [{k: v for k, v in ns.items() if k != "original_info"} for ns in new_screens]
-        write_manifest(sdir, manifest)
-        
-        accumulated_usage = empty_token_usage()
-        accumulated_estimated = False
-        
-        from concurrent.futures import ThreadPoolExecutor
-        manifest_lock = threading.Lock()
-        fatal_errors: list[str] = []
-        new_screens_info = [{"id": ns["id"], "name": ns["name"], "prompt": ns["original_info"]["prompt"]} for ns in new_screens]
-        shared_ctx = _build_shared_layout_context(new_screens_info)
-        
-        def _gen_iter_screen_worker(idx: int, ns: dict[str, Any]) -> None:
-            nonlocal accumulated_usage, accumulated_estimated
-            sid = ns["id"]
-            sname = ns["name"]
-            sprompt = ns["original_info"]["prompt"]
-            spattern = ns["original_info"].get("layout_pattern") or detect_layout_pattern(sprompt, device=device)
-            
-            # Pre-register round history entry under lock
-            with manifest_lock:
-                round_index = _next_round_index(manifest, sid)
-                rel = f"screens/{sid}_r{round_index}.html"
-                
-                round_entry = {
-                    "round_index": round_index,
-                    "screen_id": sid,
-                    "html_path": rel,
-                    "prompt": sprompt,
-                    "reasoning_content": None,
-                    "process_log": list(log[log_start:]),
-                    "at": now_iso(),
-                }
-                history_local = list(manifest.get("round_history") or [])
-                existing_idx = next((i for i, h in enumerate(history_local) if h.get("screen_id") == sid and h.get("round_index") == round_index), None)
-                if existing_idx is not None:
-                    history_local[existing_idx] = round_entry
-                else:
-                    history_local.append(round_entry)
-                manifest["round_history"] = history_local
-                
-                for s in manifest.get("screens") or []:
-                    if str(s.get("id")) == sid:
-                        s["html_path"] = rel
-                        s["active_round_index"] = round_index
-                
-                update_process_status(
-                    sdir,
-                    manifest,
-                    text=f"Generating {len(new_screens)} screens concurrently (Screen «{sname}»)...",
-                    status="generating_ui",
-                    model_id=model_id,
-                    model_name=model_name,
-                )
-                write_manifest(sdir, manifest)
-            
-            screen_spec = {**spec_dict, "layout_pattern": spattern}
-            html = ""
-            ui_reasoning = None
-            ui_fail_reason = None
-            
-            if is_model_available(router, model_id):
-                try:
-                    from src.design import service
-                    gen_ui = getattr(service, "_generate_ui_html", _generate_ui_html)
-                    html, ui_reasoning, ui_usage, ui_usage_estimated, ui_fail_reason = gen_ui(
-                        router,
-                        user_prompt=sprompt,
-                        spec=screen_spec,
-                        device=device,
-                        model_id=model_id,
-                        design_md=design_md,
-                        md_text=md_text if has_md else None,
-                        url_snapshot=url_snapshot if has_url else None,
-                        has_image=has_image,
-                        image_data_url=image_data_url if ui_vision_ok else None,
-                        current_html="",
-                        instruction="",
-                        shared_layout_context=shared_ctx,
-                    )
-                    with manifest_lock:
-                        accumulated_usage = merge_token_usage(accumulated_usage, ui_usage)
-                        accumulated_estimated = accumulated_estimated or ui_usage_estimated
-                except Exception as exc:
-                    logger.warning("design ui LLM failed for iterated screen %s run_id=%s err=%s", sid, run_id, exc)
-                    ui_fail_reason = str(exc)
-            
-            llm_available = is_model_available(router, model_id)
-            if not _html_has_visible_content(html):
-                if llm_available:
-                    detail = ui_fail_reason or "the model returned no valid HTML"
-                    err_msg = f"Generation failed for screen «{sname}» — {detail}. Please try again."
-                    with manifest_lock:
-                        fatal_errors.append(err_msg)
-                    return
-                html = _fallback_ui_html(sprompt, screen_spec, device=device)
-            
-            # Inject unified tailwind config for cross-screen style consistency
-            html = inject_unified_tailwind_config(html, spec_dict, design_md)
-            
-            # Write html and update round entry under lock
-            (sdir / "screens").mkdir(exist_ok=True)
-            (sdir / rel).write_text(html, encoding="utf-8")
-            
-            with manifest_lock:
-                history_local = list(manifest.get("round_history") or [])
-                for h in history_local:
-                    if h.get("screen_id") == sid and h.get("round_index") == round_index:
-                        h["reasoning_content"] = ui_reasoning
-                        h["process_log"] = list(log[log_start:])
-                        break
-                manifest["round_history"] = history_local
-                
-                for s in manifest.get("screens") or []:
-                    if str(s.get("id")) == sid:
-                        s["html_path"] = rel
-                        s["active_round_index"] = round_index
-                write_manifest(sdir, manifest)
-        
-        # Execute all generations concurrently
-        max_workers = min(len(new_screens), 8) if len(new_screens) > 0 else 1
-        with ThreadPoolExecutor(max_workers=max_workers) as executor:
-            list(executor.map(lambda item: _gen_iter_screen_worker(item[0], item[1]), enumerate(new_screens)))
-        
-        # If any fatal error occurred during the parallel runs, report it and exit
-        if fatal_errors:
-            err_msg = fatal_errors[0]
-            _finalize_assistant_step(
-                log,
-                text=err_msg,
-                status="error",
-                model_id=model_id,
-                model_name=model_name,
-                usage=accumulated_usage,
-                usage_estimated=accumulated_estimated,
-                replace_statuses={"generating_ui", "iterating"},
-            )
-            manifest["status"] = "error"
-            manifest["error"] = err_msg
-            manifest["process_log"] = log
-            write_manifest(sdir, manifest)
-            return public_session_payload(manifest, sdir)
-        
-        added_names = ", ".join(f"«{ns['name']}»" for ns in new_screens)
-        _finalize_assistant_step(
-            log,
-            text=f"Added {added_names} successfully. Select any screen to refine further.",
-            status="ready",
-            model_id=model_id,
-            model_name=model_name,
-            usage=accumulated_usage,
-            usage_estimated=accumulated_estimated,
-            replace_statuses={"generating_ui", "iterating"},
-        )
-        
-        # Ensure the final history entry has the full process log
-        history = list(manifest.get("round_history") or [])
-        if history:
-            history[-1] = {**history[-1], "process_log": list(log[log_start:])}
-            manifest["round_history"] = history
-            
-        manifest["last_iterate_action"] = action
-        manifest["last_iterate_screen_id"] = new_screens[-1]["id"]
-        manifest["status"] = "ready"
-        write_manifest(sdir, manifest)
-
-    elif action == "duplicate":
-        base_id = str(target_id or screens[0].get("id") or "main")
-        base = next((s for s in screens if str(s.get("id")) == base_id), screens[0])
-        base_html_path = _resolve_screen_html_path(sdir, base)
-        if not base_html_path.is_file():
-            base_html_path = sdir / "screens" / f"{base['id']}.html"
-        base_html = base_html_path.read_text(encoding="utf-8") if base_html_path.is_file() else ""
-        new_id = _next_screen_id(screens)
-        
-        round_index = _next_round_index(manifest, new_id)
-        rel = f"screens/{new_id}_r{round_index}.html"
-        
-        (sdir / "screens").mkdir(exist_ok=True)
-        (sdir / rel).write_text(base_html, encoding="utf-8")
-        
-        round_entry = {
-            "round_index": round_index,
-            "screen_id": new_id,
-            "html_path": rel,
-            "prompt": f"Duplicate of {base.get('name') or 'Screen'}",
-            "reasoning_content": None,
-            "process_log": list(log[log_start:]),
-            "at": now_iso(),
-        }
-        history = list(manifest.get("round_history") or [])
-        history.append(round_entry)
-        manifest["round_history"] = history
-        
-        new_screen = {
-            "id": new_id,
-            "name": f"Copy of {base.get('name') or 'Screen'}",
-            "position": {
-                "x": _screen_layout_x(
-                    screens,
-                    device=device,
-                    has_source=has_md or has_image or has_url,
-                ),
-                "y": _DESIGN_ROW_Y,
-            },
-            "html_path": rel,
-            "active_round_index": round_index,
-        }
-        screens.append(new_screen)
-        manifest["screens"] = screens
-        
-        _finalize_assistant_step(
-            log,
-            text=f"Duplicated screen as «{new_screen['name']}».",
-            status="ready",
-            model_id=model_id,
-            model_name=model_name,
-            usage=empty_token_usage(),
-            usage_estimated=False,
-            replace_statuses={"iterating"},
-        )
-        manifest["last_iterate_action"] = "duplicate"
-        manifest["last_iterate_screen_id"] = new_id
-        manifest["status"] = "ready"
-        write_manifest(sdir, manifest)
-    else:
-        screen_id = str(target_id or screens[0].get("id") or "main")
-        screen = next((s for s in screens if str(s.get("id")) == screen_id), screens[0])
-        screen_id = str(screen["id"])
-        html_path = _resolve_screen_html_path(sdir, screen)
-        if not html_path.is_file():
-            html_path = sdir / "screens" / f"{screen_id}.html"
-        current = html_path.read_text(encoding="utf-8") if html_path.is_file() else ""
-        html = current
-        element_hint = ""
-        if element_path or element_label:
-            element_hint = (
-                f"Focus ONLY on this element/region: {element_label or ''} "
-                f"({element_path or ''}). Keep the rest of the page intact when possible.\n"
-            )
-        merged_prompt = _merged_design_prompt(manifest, instruction)
-        device = str(manifest.get("device") or "web")
-        spec_dict = spec if isinstance(spec, dict) else _fallback_spec(merged_prompt)
-        ui_reasoning: str | None = None
-        ui_usage = empty_token_usage()
-        ui_usage_estimated = False
-        if is_model_available(router, model_id):
-            try:
-                from src.design import service
-                gen_ui = getattr(service, "_generate_ui_html", _generate_ui_html)
-                candidate, ui_reasoning, ui_usage, ui_usage_estimated, _ui_fail = gen_ui(
-                    router,
-                    user_prompt=merged_prompt,
-                    spec=spec_dict,
-                    device=device,
-                    model_id=model_id,
-                    design_md=design_md,
-                    md_text=md_text if has_md else None,
-                    url_snapshot=url_snapshot if has_url else None,
-                    has_image=False,
-                    image_data_url=None,
-                    current_html=current,
-                    instruction=f"{instruction}\n{element_hint}",
-                    fallback_html=current,
-                )
-                if _html_has_visible_content(candidate) and not _html_essentially_same(
-                    candidate, current
-                ):
-                    # Guard: reject if page type completely changed (dashboard→login etc.)
-                    if current and _detect_html_intent(candidate) != _detect_html_intent(current):
-                        logger.warning(
-                            "design iterate modify type-changed run_id=%s — keeping current HTML",
-                            run_id,
-                        )
-                    else:
-                        html = candidate
-                elif not is_model_available(router, model_id):
-                    html = _fallback_ui_html(merged_prompt, spec_dict, device=device)
-                else:
-                    logger.warning(
-                        "design iterate modify unchanged/blank run_id=%s — keeping current HTML",
-                        run_id,
-                    )
-            except Exception as exc:
-                logger.warning("design iterate modify failed: %s", exc)
-                if not is_model_available(router, model_id):
-                    html = _fallback_ui_html(merged_prompt, spec_dict, device=device)
-        else:
-            html = _fallback_ui_html(merged_prompt, spec_dict, device=device)
-        if not is_model_available(router, model_id) and (
-            not _html_has_visible_content(html) or _html_essentially_same(html, current)
-        ):
-            html = _fallback_ui_html(merged_prompt, spec_dict, device=device)
-        rel = screen.get("html_path") or f"screens/{screen_id}_r0.html"
-        (sdir / "screens").mkdir(exist_ok=True)
-        # Inject unified tailwind config to maintain style consistency after modification
-        html = inject_unified_tailwind_config(html, spec, design_md)
-        (sdir / rel).write_text(html, encoding="utf-8")
-        screen["html_path"] = rel
-        iterate_ready_text = (
-            f"Updated the artboard — wrote {rel}. What else?"
-        )
-        if not is_model_available(router, model_id):
-            iterate_ready_text += f"\n\n⚠️ Warning: Model '{model_id}' is not available (API key missing in Settings -> Models). Using offline fallback templates."
-        _finalize_assistant_step(
-            log,
-            text=iterate_ready_text,
-            status="ready",
-            model_id=model_id,
-            model_name=model_name,
-            usage=ui_usage,
-            usage_estimated=ui_usage_estimated,
-            replace_statuses={"iterating"},
-        )
-        manifest["last_iterate_action"] = "modify"
-        manifest["last_iterate_screen_id"] = screen_id
-
-    clear_fake_thumbnail(sdir)
-    manifest.pop("thumbnail", None)
-    manifest["process_log"] = log
-    manifest["status"] = "ready"
-    write_manifest(sdir, manifest)
-    return public_session_payload(manifest, sdir)
-
-
-def delete_screen(
-    run_id: str,
-    screen_id: str,
-) -> dict[str, Any]:
-    sdir = session_dir(run_id)
-    manifest = read_manifest(sdir)
-    screens = list(manifest.get("screens") or [])
-    target = next((s for s in screens if str(s.get("id")) == screen_id), None)
-    if not target:
-        raise DesignError(f"Screen {screen_id} not found")
-    if target.get("deleted"):
-        raise DesignError(f"Screen {screen_id} is already deleted")
-    removed_name = str(target.get("name") or screen_id)
-    round_index = _next_round_index(manifest, screen_id)
-    for s in screens:
-        if str(s.get("id")) == screen_id:
-            s["deleted"] = True
-            s["deleted_at_round"] = round_index
-            break
-
-    instruction = f"Delete screen: {removed_name}"
-    log = list(manifest.get("process_log") or [])
-    log_start = len(log)
-    log.append({"role": "user", "text": instruction, "at": now_iso()})
-    log.append({"role": "assistant", "text": f"Removed «{removed_name}» from the canvas.", "status": "ready", "at": now_iso()})
-    history = list(manifest.get("round_history") or [])
-    history.append({
-        "round_index": round_index,
-        "screen_id": screen_id,
-        "html_path": f"screens/{screen_id}_r{round_index}.html",
-        "prompt": instruction,
-        "reasoning_content": None,
-        "process_log": list(log[log_start:]),
-        "at": now_iso(),
-        "action": "delete",
-    })
-    manifest["round_history"] = history
-    manifest["process_log"] = log
-    manifest["status"] = "ready"
-    manifest["last_iterate_action"] = "delete"
-    manifest["last_iterate_screen_id"] = None
-    write_manifest(sdir, manifest)
-    return public_session_payload(manifest, sdir)
